% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addEcopathParams}
\alias{addEcopathParams}
\title{Add Ecopath parameters to species parameters}
\usage{
addEcopathParams(species_params, ecopath_params, species_to_groups = list())
}
\arguments{
\item{species_params}{A data frame with mizer species parameters}

\item{ecopath_params}{A data frame with Ecopath parameters for each group
as exported by the Ecopath software.}

\item{species_to_groups}{A named list where the names are mizer species and
the values are vectors of Ecopath groups, see Details below.}
}
\value{
The mizer species parameter data frame with the added columns
\code{biomass_observed}, \code{consumption_observed}, \code{production_observed} and
\code{ecopath_groups}.
}
\description{
Determines the biomass, consumption and production rates for each species in
the ecopath model and adds these to the species parameters. Ecopath works
with "groups" where mizer works with "species". The \code{species_to_groups}
parameter is a named list that maps Ecopath groups to mizer species. The
names must be the same as the species in \code{species_params} and the values must
be the names of the corresponding Ecopath groups. Species in the mizer model
that are not mapped to groups in the Ecopath model are left unchanged.
}
\details{
In case the Ecopath model has groups that are split into stanzas, then the
value in \code{species_to_groups} should be a vector with the names of the stanzas
that need to be combined. The biomass, production and consumption of these
stanzas are added together to give the values for the corresponding mizer
species.

The species mapping from \code{species_to_groups} is stored in the \code{ecopath_groups}
column of the species_params data frame. This column is a list column so that
it can store a vector of groups in the case where a species is made up of
several groups.

The biomasses are added to the species_params data frame in the
\code{biomass_observed} column. The consumption rates are put into a
\code{consumption_observed} column and the production rates are put into a
\code{production_observed} column. The names of the Ecopath groups associated to
each mizer species are put into the \code{ecopath_groups} column. This column is a
list column so that it can store a vector of groups in the case where a
species is made up of several groups.

If \code{biomass_observed}, \code{production_observed} or \code{consumption_observed} columns
already have values differing from the Ecopath values, a warning is issued
and the values are overwritten.

If a species in \code{species_params} is not found in \code{species_to_groups}, a
message is issued and the species is skipped.
}
