% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass_discard_rates}
\alias{biomass_discard_rates}
\title{Biomass discard rates}
\format{
A grouped data frame with columns:
\describe{
\item{gear_category}{Gear category}
\item{species}{Species name}
\item{country}{Country}
\item{target_assemblage}{Target assemblage}
\item{total_landings}{Total landings (t)}
\item{tot_discards_tonnes}{Total discards (t)}
\item{discard_rate}{Discard rate (proportion)}
}
}
\usage{
biomass_discard_rates
}
\description{
Discard rates by gear category and species, derived from STECF data.
}
\keyword{datasets}
