% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchSelectivity.R
\name{catchSelectivity}
\alias{catchSelectivity}
\title{Calculate catch selectivity params from catch data}
\usage{
catchSelectivity(params, catch)
}
\arguments{
\item{params}{A MizerParams object. a and b species parameters are required
in the params object}

\item{catch}{A data frame containing at least the following columns:
'species' (common names for each species), 'gear', 'length', and 'catch'
(scaled to total g/m^2/year). Additional value columns are allowed.}
}
\value{
A MizerParams object with selectivity for catch data
}
\description{
This function determines the selectivity parameter for the params object from
the user provided scaled catch data. Catch data is rebinned into mizer bins
and then divided by the total abundace. When abundance reaches 0 this means
selectivity values will reach Inf and a warning will be issued that the
steady state model does not support large enough fish according to the catch
data.
}
