% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celtic_params}
\alias{celtic_params}
\title{Celtic Sea MizerParams object}
\format{
A MizerParams object with 12 species.
}
\usage{
celtic_params
}
\description{
A 12-species MizerParams object for the Celtic Sea, built from the Ecopath
model of Lauria et al. (2016). Used as the main example object in tests and
vignettes.
}
\keyword{datasets}
