% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDiet.R
\name{checkDietMatrix}
\alias{checkDietMatrix}
\title{Check if a diet matrix is valid}
\usage{
checkDietMatrix(diet_matrix)
}
\arguments{
\item{diet_matrix}{The diet matrix to check}
}
\value{
NULL invisibly. Throws an error if any check fails.
}
\description{
This function performs validation checks on a diet matrix to ensure it meets
the requirements for use in mizer.
}
\keyword{internal}
