% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnergeticComponents.R
\name{computeEnergeticComponents}
\alias{computeEnergeticComponents}
\title{Set up intake rate and metabolic loss parameters for a tuned non-interacting
model. Take the existing tuned encounter rates and compute the correct
internal component values, and sets them based on a supplied assimilation
efficiency, critical feeding level,  and feeding level. Defaults for these
are set following Andersen (2019).}
\usage{
computeEnergeticComponents(params, fc = 0.2, f = 0.6)
}
\arguments{
\item{params}{A MizerParams object}

\item{fc}{Critical feeding level. Default is 0.2}

\item{f}{Feeding level. Default is 0.6}
}
\value{
A MizerParams object with the adjusted max intake rate (h) and
metablolic loss (k) parameters
}
\description{
Set up intake rate and metabolic loss parameters for a tuned non-interacting
model. Take the existing tuned encounter rates and compute the correct
internal component values, and sets them based on a supplied assimilation
efficiency, critical feeding level,  and feeding level. Defaults for these
are set following Andersen (2019).
}
