% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{control_description}
\alias{control_description}
\title{Return the description for the control}
\usage{
control_description(control)
}
\arguments{
\item{control}{The control name}
}
\value{
The control description string, or \code{NULL} if none is defined
}
\description{
This function returns the description for the control, if the control author
has defined one in a variable \verb{<control>ControlDescription}. The description
is used as a tooltip on the control's title and on the link to the control in
the sidebar. If no description is defined, \code{NULL} is returned and no tooltip
is shown.
}
\keyword{internal}
