% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{control_title}
\alias{control_title}
\title{Return the title for the control}
\usage{
control_title(control)
}
\arguments{
\item{control}{The control name}
}
\value{
The control title
}
\description{
This function is used to return the title for the control. This is either
defined by the control author or otherwise is the control name supplied by
the user.
}
\keyword{internal}
