% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_calculated_histories}
\alias{cs_calculated_histories}
\title{Celtic Sea calculated life history parameters}
\format{
A data frame with columns:
\describe{
\item{Scientific_name}{Scientific name}
\item{l_mat}{Length at maturity (cm)}
\item{age_mat}{Age at maturity (years)}
}
}
\usage{
cs_calculated_histories
}
\description{
Estimated length at maturity and age at maturity for Celtic Sea species,
derived from survey data.
}
\keyword{datasets}
