% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diet_matrix_example}
\alias{diet_matrix_example}
\title{Example reduced diet matrix for mizer}
\format{
A matrix with species as both row and column names, giving the proportion of each prey
in each predator’s diet.
}
\description{
The output of applying \code{reduceEcopathDiet()} to the Celtic Sea Ecopath diet matrix.
Rows and columns are matched to \code{species_params_example$species}.
}
\keyword{datasets}
