% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diets}
\alias{diets}
\title{Celtic Sea Ecopath diet matrix}
\format{
A data frame with 75 rows and 66 columns. Row names are prey
groups; column names are predator groups.
}
\usage{
diets
}
\description{
The Ecopath diet composition matrix for the Celtic Sea model of
Lauria et al. (2016), including juvenile and adult stanzas.
}
\keyword{datasets}
