% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionControl.R
\name{diffusionControl}
\alias{diffusionControl}
\alias{diffusionControlUI}
\title{Controlling the diffusion rate}
\usage{
diffusionControl(input, output, session, params, params_old, flags, ...)

diffusionControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the exponents
}
\description{
Controlling the diffusion rate
}
