% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecopath_diet_example}
\alias{ecopath_diet_example}
\title{Example Ecopath diet matrix}
\format{
A data frame with predator groups as row names and prey groups as columns.
}
\usage{
ecopath_diet_example
}
\description{
A group-level diet composition matrix derived from Lauria et al. (2016),
used as input to \code{reduceEcopathDiet()}. Columns and rows include juvenile and adult stanzas.
}
\keyword{datasets}
