% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathMatchControl.R
\name{ecopathMatchControl}
\alias{ecopathMatchControl}
\alias{ecopathMatchControlUI}
\title{Ecopath-specific match control}
\usage{
ecopathMatchControl(input, output, session, params, params_old, flags, ...)

ecopathMatchControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the matching strength
}
\description{
This control adjusts the matching strength for production and yield.
These parameters determine how strongly the model is attracted to
its Ecopath-aligned targets during the tuning process.
}
\seealso{
Other gadget controls:
\code{\link[=abundanceControl]{abundanceControl()}},
\code{\link[=ecopathDiffusionControl]{ecopathDiffusionControl()}},
\code{\link[=ecopathExponentControl]{ecopathExponentControl()}},
\code{\link[=ecopathFishingControl]{ecopathFishingControl()}},
\code{\link[=ecopathGrowthControl]{ecopathGrowthControl()}},
\code{\link[=ecopathOtherControl]{ecopathOtherControl()}},
\code{\link[=ecopathReproductionControl]{ecopathReproductionControl()}},
\code{\link[=fishingControl]{fishingControl()}},
\code{\link[=growthControl]{growthControl()}},
\code{\link[=interactionControl]{interactionControl()}},
\code{\link[=otherControl]{otherControl()}},
\code{\link[=predationControl]{predationControl()}},
\code{\link[=reproductionControl]{reproductionControl()}},
\code{\link[=resourceControl]{resourceControl()}}
}
\concept{gadget controls}
