% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathSpectraTab.R
\name{ecopathSpectraTab}
\alias{ecopathSpectraTab}
\alias{ecopathSpectraTabUI}
\title{Ecopath spectra tab}
\usage{
ecopathSpectraTab(input, output, session, params, logs, trigger_update, ...)

ecopathSpectraTabUI(params, help = TRUE, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering updates of species
parameter controls.}

\item{...}{Unused}

\item{help}{Boolean. If \code{FALSE}, omit the explanatory text from the tab.}
}
\description{
This tab displays the size spectra of the forground and background
species. It provides controls for:
\itemize{
\item \strong{Scaling the background}: Adjusting the resource and background
species' biomass.
\item \strong{Adjusting/removing background species}: Fine-tuning the
background community to align with the foreground.
}
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=exampleTab]{exampleTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
