% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{error_fun}
\alias{error_fun}
\title{Error function}
\usage{
error_fun(e)
}
\arguments{
\item{e}{The error object}
}
\description{
This function is called when an error occurs. It shows a modal dialog with the error message.
}
