% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTab.R
\name{exampleTab}
\alias{exampleTab}
\title{Example tab for extending the tuning gadget}
\usage{
exampleTab(input, output, session, params, logs, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
A minimal server function demonstrating how to add a custom tab to
\code{\link[=tuningGadget]{tuningGadget()}}. It plots mortality for the selected species as either a
proportion or a rate.
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
