% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{finalise_params}
\alias{finalise_params}
\title{Finalise the params object after tuning}
\usage{
finalise_params(p)
}
\arguments{
\item{p}{The params object}
}
\value{
The finalised params object
}
\description{
This function is called when a params object is downloaded or when the done
button is pressed. It clears the undo attribute and restores the
Beverton-Holt reproduction settings.
}
\keyword{internal}
