% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishing_deaths}
\alias{fishing_deaths}
\title{Fishing deaths data frame}
\format{
A data frame with columns:
\describe{
\item{species}{Species name}
\item{total_weight_dead_gear_per_area}{Total weight of dead fish per
gear per area (t/km^2/year)}
\item{gear}{Gear category}
}
}
\usage{
fishing_deaths
}
\description{
Total weight of dead discards per gear per area, used in discard rate
calculations.
}
\keyword{datasets}
