% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiffusion.R
\name{getDiffusion}
\alias{getDiffusion}
\title{Get diffusion rate}
\usage{
getDiffusion(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  t = 0,
  order = 2,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{order}{By setting this to integer values above 2 one can get the
coefficients of higher-order terms in the expansion of the jump-growth
equation. The default is 2, which gives the diffusion rate.}

\item{...}{Unused}
}
\value{
A named two dimensional array (predator species x predator size) with
the diffusion rates.
}
\description{
The diffusion rate \eqn{d_i(w)} for species \eqn{i}
and weight \eqn{w} has contributions from the encounter of
fish prey and of resource. This is determined by summing over all prey
species and the resource spectrum and then integrating over all prey sizes
\eqn{w_p}, weighted by predation kernel \eqn{\phi(w,w_p)}:
\deqn{
d_i(w) = (1-f_i(w))(\alpha_i(1-\psi_i(w)))^2\gamma_i(w) \int
\left( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) \right)
\phi_i(w,w_p) w_p^2 \, dw_p.
}{(1-f_i(w))(\alpha_i(1-\psi_i(w)))^2\gamma_i(w) \int
( \theta_{ip} N_R(w_p) + \sum_{j} \theta_{ij} N_j(w_p) )
\phi_i(w,w_p) w_p^2 dw_p.}
Here \eqn{N_j(w)} is the abundance density of species \eqn{j} and
\eqn{N_R(w)} is the abundance density of resource.
The overall prefactor \eqn{\gamma_i(w)} determines the predation power of the
predator. It could be interpreted as a search volume and is set with the
\code{\link[=setSearchVolume]{setSearchVolume()}} function. The predation kernel
\eqn{\phi(w,w_p)} is set with the \code{\link[=setPredKernel]{setPredKernel()}} function. The
species interaction matrix \eqn{\theta_{ij}} is set with \code{\link[=setInteraction]{setInteraction()}}
and the resource interaction vector \eqn{\theta_{ip}} is taken from the
\code{interaction_resource} column in \code{params@species_params}.
\eqn{f(w)} is the feeding level calculated with
\code{\link[=getFeedingLevel]{getFeedingLevel()}}. \eqn{\psi(w)} is the proportion of the available energy
that is invested in reproduction instead of growth, as stored in \code{params@psi}.
}
\details{
The function returns values also for sizes outside the size-range of the
species. These values should not be used, as they are meaningless.
}
\concept{mizer rate functions}
