% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getOffspringProduction}
\alias{getOffspringProduction}
\title{Get rate at which offspring biomass is produced}
\usage{
getOffspringProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of offspring biomass production for each species
}
\description{
For each species, returns the rate at which offspring biomass is produced is
the product of the rate of offspring production (obtained with \code{\link[mizer:getRDD]{mizer::getRDD()}})
and the offspring biomass \eqn{w_{0.i}}.
}
\examples{
getOffspringProduction(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveEfficiency]{getReproductiveEfficiency()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
