% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthTab.R
\name{growthTab}
\alias{growthTab}
\alias{growthTabUI}
\title{Growth tab for tuning gadget
inheritParams biomassTab}
\usage{
growthTab(input, output, session, params, logs, size_at_age = NULL, ...)

growthTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{size_at_age}{A data frame with observed size at age data to be plotted
on top of growth curves. Should contain columns \code{species} (species name as
used in the model), \code{age} (in years) and either \code{weight} (in grams) or
\code{length} (in cm). If both \code{weight} and \code{length} are provided, only \code{weight}
is used.}

\item{...}{Unused}
}
\description{
Growth tab for tuning gadget
inheritParams biomassTab
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=exampleTab]{exampleTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
