% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAllometricParams.R
\name{isAllometric}
\alias{isAllometric}
\title{Test whether a model has allometric encounter and mortality rates}
\usage{
isAllometric(params, tol = 1e-05)
}
\arguments{
\item{params}{A MizerParams object}

\item{tol}{The relative tolerance.}
}
\value{
TRUE if the model has allometric encounter and mortality rates
}
\description{
This function returns TRUE when the model has allometric encounter and
mortality rates and FALSE otherwise.
}
