% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInteractive.R
\name{makeInteractive}
\alias{makeInteractive}
\title{Make species interact with given interaction matrix}
\usage{
makeInteractive(params, interaction)
}
\arguments{
\item{params}{A MizerParams object}

\item{interaction}{An interaction matrix where the entry in row i and
column j determines how strongly species i predates on species j.}
}
\value{
The modified MizerParams object
}
\description{
This function should be called with a non-interacting model (where there is
only external mortality and external encounter) and an interaction matrix.
The external encounter and mortality rates will then be decreased by the
amount of the predation encounter and mortality that is created by that
interaction matrix.
}
