% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchConsumption.R
\name{matchConsumption}
\alias{matchConsumption}
\title{Match the model's total consumption to observed estimates}
\usage{
matchConsumption(params, species = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{A vector of species names or indices. If NULL,
applies to all species with a provided \code{consumption_observed}.}
}
\value{
A \code{MizerParams} object with adjusted encounter and metabolic
respiration rates.
}
\description{
This function sets the metabolic respiration rate so that the total
consumption matches the parameter \code{consumption_observed} for each species.
The function preserves the energy available for growth and reproduction by
adjusting external encounter rates to compensate for changes in metabolic
loss.
}
\details{
To adjust both the encounter rate and the metabolic respiration rate the
function assumes that they are given by power laws with the same exponent
\code{n}, so it sets the species parameter \code{p} to the same value as \code{n}. A
warning is issued if the exponent \code{p} had to be changed for any species.

Any of the selected species for which \code{consumption_observed} is NA will be
quietly ignored.

If the resulting total metabolic respiration would be less than 10\% of
total production, the function clamps it to 10\% of total production and
issues a warning. In that case the consumption will not perfectly match
\code{consumption_observed} for those species.

The energy available for growth and reproduction is always preserved, so
the steady state spectra are unchanged.
}
\examples{
params <- suppressWarnings(matchConsumption(celtic_params))
# The energy available for growth and reproduction is not changed
all.equal(getEReproAndGrowth(params),
          getEReproAndGrowth(celtic_params))
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchDiet}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchProduction}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
