% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDiet.R
\name{matchDiet}
\alias{matchDiet}
\title{Match the diet matrix of the model to an observed diet matrix}
\usage{
matchDiet(params, diet_matrix, min_w_pred = 0, max_w_pred = Inf)
}
\arguments{
\item{params}{A MizerParams object}

\item{diet_matrix}{The Ecopath diet matrix as produced by
\code{reduceEcopathDiet()}.}

\item{min_w_pred}{The minimum weight of predators to include in the diet
matrix. Default is 0.}

\item{max_w_pred}{The maximum weight of predators to include in the diet
matrix. Default is \code{Inf}.}
}
\value{
A MizerParams object with the interaction matrix set so that the
desired diet is achieved.
}
\description{
This function adjusts the interaction matrix of the model so that its diet
matrix matches the given diet matrix (typically derived using\code{reduceEcopathDiet()}).
}
\details{
This function first converts the input diet matrix from proportions to
absolute consumption using species-specific consumption rates. It then
calculates the interaction strengths required to reproduce this diet and
applies them using \code{makeInteracting()}. External mortality and encounter
terms are adjusted to preserve the model’s steady state.

If \code{min_w_pred} or \code{max_w_pred} are provided, the diet matrix is assumed to
represent consumption by individuals within that size range

If matching the diet requires interaction strengths that would imply
negative external mortality or encounter rates for any species, warnings
will be issued by \code{makeInteracting()}. These indicate that the model
assumptions may be biologically unrealistic or that the input diet matrix
is incompatible with other model parameters.
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchConsumption}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchProduction}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
