% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{matchEcopath}
\alias{matchEcopath}
\alias{isEcopathMatched}
\title{Match the model to the Ecopath data}
\usage{
matchEcopath(params, tol = 0.1, max_iter = 10)

isEcopathMatched(params, tol = 0.1)
}
\arguments{
\item{params}{A MizerParams object}

\item{tol}{The relative tolerance for the match. Default is 0.1.}

\item{max_iter}{The maximum number of iterations. Default is 100.}
}
\value{
A MizerParams object matched to the Ecopath data
}
\description{
This function adjusts the model parameters to match the Ecopath data. It does
this by adjusting the consumption, production, and yield parameters.
}
\details{
The \code{match} function calls \code{matchConsumption} and
\code{matchProduction} repeatedly until both consumption and production
matches within the tolerance specified by the \code{tol} parameter. The function
will return a warning if the maximum number of iterations is reached without
converging.
}
