% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGonadicProportion.R
\name{matchGonadicProportionOnce}
\alias{matchGonadicProportionOnce}
\title{Match the gonadic proportion of a species to a given value.}
\usage{
matchGonadicProportionOnce(params, steady = TRUE)
}
\arguments{
\item{params}{A MizerParams object}

\item{steady}{Whether to return the model to a steady state after adjusting
the gonadic proportion. Default is TRUE.}
}
\value{
A MizerParams object with the gonadic proportion matched
}
\description{
This function adjusts the \code{w_repro_max} parameters to match the gonadic
proportion of the species.
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchConsumption}()},
\code{\link{matchDiet}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchProduction}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
