% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRespiration.R
\name{matchRespirationOnce}
\alias{matchRespirationOnce}
\title{Match the respiration rate to that required by the ecopath production rate.}
\usage{
matchRespirationOnce(params, steady = TRUE)
}
\arguments{
\item{params}{A MizerParams object}

\item{steady}{Whether to return the model to a steady state after adjusting
the respiration. Default is TRUE.}
}
\value{
A MizerParams object with the respiration matched
}
\description{
This function sets the respiration rate (called the metabolic rate in mizer)
according to the first master equation of Ecopath, which is:
\deqn{Q = P + R + U}
where \eqn{Q} is the consumption rate, \eqn{P} is the production rate,
\eqn{R} is the respiration rate, and \eqn{U} is the unassimilated part of the
consumption rate, i.e., \eqn{U = (1 - \alpha) Q}, where \eqn{\alpha} is the
assimilation efficiency. Solving this for \eqn{R} we get:
\deqn{R = Q \alpha - P}
}
\details{
The production rate of a species is the sum of the somatic production rate and
the gonad production rate. The somatic production rate is the production rate

The Ecopath production rate is the somatic production rate, i.e., the production
rate of the somatic tissue. To get the total production rate we need to divide
the somatic production rate by \eqn{1 - f}, where \eqn{f} is the gonad
proportion. This is because the gonad production is not included in the Ecopath
production rate.

This function calculates \eqn{Q} using \code{getConsumption()}
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchConsumption}()},
\code{\link{matchDiet}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchProduction}()}
}
\concept{match functions}
