% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerEcopath-package.R
\docType{package}
\name{mizerEcopath-package}
\alias{mizerEcopath}
\alias{mizerEcopath-package}
\title{mizerEcopath: Building a Mizer version of an Ecopath model}
\description{
An interactive Shiny gadget for tuning a mizer model built with mizerFromEcopath, together with a growth-diffusion and age-at-size module and an allometric alternative to the von Bertalanffy starting model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gustavdelius/mizerEcopath}
  \item Report bugs at \url{https://github.com/gustavdelius/mizerEcopath/issues}
}

}
\author{
\strong{Maintainer}: Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]
}

}
\keyword{internal}
