% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newAllometricParams.R
\name{newAllometricParams}
\alias{newAllometricParams}
\title{Set up parameters for a model with allometric encounter and mortality rates}
\usage{
newAllometricParams(species_params, no_w = 200)
}
\arguments{
\item{species_params}{A data frame with species parameters}

\item{no_w}{The number of weight bins to use in the model}
}
\value{
A MizerParams object
}
\description{
This function sets up a model with allometric encounter and mortality rates
for each species and no interaction between the species or with the resource.
The initial abundances are set to the steady-state abundances with the total
biomass for each species matched to observations.
The metabolic respiration rate, the feeding level and the reproduction level
are set to zero.
}
\details{
If the exponent n of the power-law encounter rate is not provided as a species
parameter, it is set to 0.7. The exponent d of the power-law mortality rate is
\emph{always} imposed as n - 1, regardless of any supplied value.

If the species parameter \code{alpha}, which gives the proportion of the consumption
that is assimilated, is not provided, it is set to 0.8, the default value
used by Ecopath.

The coefficient of the power-law encounter rate for each species is chosen so
that the species grows to maturity size by its maturity age. The coefficient
of the power-law mortality rate is computed from the growth rate and is chosen so that the juvenile biomass density has a
slope of -0.2.

The species uses \code{matchGrowth()} to adjust the encounter rate coefficient to
produce the desired growth rate. It uses \code{matchBiomasses()} to match the
biomass to the observations and \code{steadySingleSpecies()} to set the model to
steady state.

Because the model does not make use of the resource spectrum, the resource
dynamics is switched off.
}
