% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_3_spp_model_final}
\alias{ns_3_spp_model_final}
\title{North Sea 3-species model: final calibrated state}
\format{
A MizerParams object with 3 species (Cod, Haddock, Sprat).
}
\usage{
ns_3_spp_model_final
}
\description{
A 3-species MizerParams object for the North Sea after full calibration.
}
\keyword{datasets}
