% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_3_spp_model_initial}
\alias{ns_3_spp_model_initial}
\title{North Sea 3-species model: initial state}
\format{
A MizerParams object with 3 species (Cod, Haddock, Sprat).
}
\usage{
ns_3_spp_model_initial
}
\description{
A 3-species MizerParams object for the North Sea before calibration.
}
\keyword{datasets}
