% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_3_spp_model_trial}
\alias{ns_3_spp_model_trial}
\title{North Sea 3-species model: trial state}
\format{
A MizerParams object with 3 species (Cod, Haddock, Sprat).
}
\usage{
ns_3_spp_model_trial
}
\description{
A 3-species MizerParams object for the North Sea at an intermediate
calibration stage.
}
\keyword{datasets}
