% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherControl.R
\name{otherControl}
\alias{otherControl}
\alias{otherControlUI}
\title{Other parameters control for tuning gadget}
\usage{
otherControl(input, output, session, params, params_old, flags, ...)

otherControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control provides sliders for species parameters that do not fit into
the other categories. It includes:
\itemize{
\item \strong{ks and p}: Metabolic rate coefficient and exponent. Changing \code{p}
automatically adjusts \code{ks} to keep the metabolic rate at maturity
size constant.
\item \strong{k}: Activity rate coefficient.
\item \strong{z_ext}: External mortality at 1g. Changing this scales
the external mortality at all other sizes by the same factor.
\item \strong{alpha}: Assimilation efficiency.
}
}
\seealso{
Other gadget controls:
\code{\link[=abundanceControl]{abundanceControl()}},
\code{\link[=ecopathDiffusionControl]{ecopathDiffusionControl()}},
\code{\link[=ecopathExponentControl]{ecopathExponentControl()}},
\code{\link[=ecopathFishingControl]{ecopathFishingControl()}},
\code{\link[=ecopathGrowthControl]{ecopathGrowthControl()}},
\code{\link[=ecopathOtherControl]{ecopathOtherControl()}},
\code{\link[=ecopathReproductionControl]{ecopathReproductionControl()}},
\code{\link[=fishingControl]{fishingControl()}},
\code{\link[=growthControl]{growthControl()}},
\code{\link[=interactionControl]{interactionControl()}},
\code{\link[=predationControl]{predationControl()}},
\code{\link[=reproductionControl]{reproductionControl()}},
\code{\link[=resourceControl]{resourceControl()}}
}
\concept{gadget controls}
