% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsumptionVsSpecies.R
\name{plotConsumptionVsSpecies}
\alias{plotConsumptionVsSpecies}
\title{Plot the Consumption against species}
\usage{
plotConsumptionVsSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Plot the Consumption against species
}
