% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProductionVsSpecies.R
\name{plotProductionVsSpecies}
\alias{plotProductionVsSpecies}
\title{Plot the production against species}
\usage{
plotProductionVsSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Plot the production against species
}
