% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_diffusion.R
\name{project_diffusion}
\alias{project_diffusion}
\title{Project forward in time with diffusion}
\usage{
project_diffusion(params, species, dt = 0.05, nsteps = 200)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species for which to solve the steady-state equation.}

\item{dt}{Time step size.}

\item{nsteps}{Number of time steps to compute.}
}
\value{
A matrix where each column corresponds to the solution at a time step.
}
\description{
This function solves the PDE for a single species
including diffusion. It uses an upwind difference scheme.
}
\details{
For more details  see the vignette
\code{vignette("diffusion")}.
}
