% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproTab.R
\name{reproTab}
\alias{reproTab}
\alias{reproTabUI}
\title{Reproduction tab for tuning gadget}
\usage{
reproTab(input, output, session, params, logs, ...)

reproTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
This tab displays parameters related to species maturity and
reproductive success. It includes:
\itemize{
\item \strong{Reproductive efficiency plot}: Shows which fraction of the energy
that the species invests into reproduction ends up as egg biomass,
both before and after density dependence. Values above the red line
at 1 are energetically impossible. See
\code{\link[=plotReproductiveEfficiency]{plotReproductiveEfficiency()}}.
\item \strong{Maturity ogive plot}: Shows the proportion of energy allocated
to reproduction (psi) vs size for the selected species.
}
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
