% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceControl.R
\name{resourceControl}
\alias{resourceControl}
\alias{resourceControlUI}
\title{Controlling the resource parameters in the tuning gadget}
\usage{
resourceControl(input, output, session, params, flags, ...)

resourceControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows the user to adjust the global resource parameters:
\itemize{
\item \strong{Sheldon exponent 'lambda'}: The slope of the resource spectrum.
\item \strong{Resource coefficient 'kappa'}: The overall abundance of
the resource.
\item \strong{Largest resource}: The size cutoff for the resource spectrum.
}
}
\seealso{
Other gadget controls:
\code{\link[=abundanceControl]{abundanceControl()}},
\code{\link[=ecopathDiffusionControl]{ecopathDiffusionControl()}},
\code{\link[=ecopathExponentControl]{ecopathExponentControl()}},
\code{\link[=ecopathFishingControl]{ecopathFishingControl()}},
\code{\link[=ecopathGrowthControl]{ecopathGrowthControl()}},
\code{\link[=ecopathOtherControl]{ecopathOtherControl()}},
\code{\link[=ecopathReproductionControl]{ecopathReproductionControl()}},
\code{\link[=fishingControl]{fishingControl()}},
\code{\link[=growthControl]{growthControl()}},
\code{\link[=interactionControl]{interactionControl()}},
\code{\link[=otherControl]{otherControl()}},
\code{\link[=predationControl]{predationControl()}},
\code{\link[=reproductionControl]{reproductionControl()}}
}
\concept{gadget controls}
