% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFeedingLevel.R
\name{setFeedingLevel}
\alias{setFeedingLevel}
\title{Set constant feeding level for each species in a non-interacting model}
\usage{
setFeedingLevel(params, feeding_level)
}
\arguments{
\item{params}{A MizerParams object}

\item{feeding_level}{The feeding level to set for each species. Optional. A
vector with one value for each species or a single value that is used for
all species. If the argument is missing, the feeding level is taken from
the \code{f0} column in the species_params object. If \code{f0} is not present, an
error is thrown.}
}
\value{
A MizerParams object with the feeding level set
}
\description{
This function adjusts the maximum intake rate in order to achieve the desired
feeding level for each species. It also adjust the external encounter rate
so that the consumption rate does not change in spite of the change of
feeding level.
}
\details{
This function only works for models where all encounter is external
encounter. Try calling \code{makeNoninteracting()} first.
}
