% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateAge.R
\name{simulateCohort}
\alias{simulateCohort}
\title{Simulate cohort evolution and calculate matrix G}
\usage{
simulateCohort(params, species, l_survey, t_max, dt = 0.05)
}
\arguments{
\item{params}{A MizerParams object.}

\item{species}{Name of the species to simulate.}

\item{l_survey}{Survey length‑bin edges (cm)}

\item{t_max}{Maximum time for simulation (years).}

\item{dt}{Time step for the model simulation (years). Default is 0.05.}
}
\value{
A list containing:
\itemize{
\item G: Matrix representing cohort size distribution over time
\item age: Age grid \link{years}
}
}
\description{
This function simulates the evolution of a cohort over time and calculates
the matrix G which represents the cohort's size distribution over time.
}
