% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_diffusion.R
\name{solve_diffusion_pde}
\alias{solve_diffusion_pde}
\title{Function to solve a reaction-diffusion-advection PDE}
\usage{
solve_diffusion_pde(d, g, mu, n_init, h, dt, nsteps)
}
\arguments{
\item{d}{A vector of diffusion coefficients.}

\item{g}{A vector of advection coefficients.}

\item{mu}{A vector of reaction coefficients.}

\item{n_init}{Initial condition for the solution.}

\item{h}{Spatial step size.}

\item{dt}{Time step size.}

\item{nsteps}{Number of time steps to compute.}
}
\value{
A matrix (time x size) with the solution of the PDE.
}
\description{
This function uses the implicit Euler method to solve a PDE of the form:
\deqn{\dot{n}(x,t) = (d(x) n'(x,t))'-(g(x)n(x,t))'-\mu(x)n(x,t)}
}
