% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_sweep.R
\name{solve_double_sweep}
\alias{solve_double_sweep}
\title{Helper function for double sweep method (Thomas algorithm)}
\usage{
solve_double_sweep(U, L, D, b)
}
\arguments{
\item{U}{Upper diagonal entries. Even though the matrix only uses
\eqn{U_1\dots U_{N-1}}, you need to supply a full vector of length \eqn{N}.
The last element will be ignored.}

\item{L}{Lower diagonal entries. Even though the matrix only uses
\eqn{L_2\dots L_N}, you need to supply a full vector of length \eqn{N}.
The first element will be ignored.}

\item{D}{Main diagonal entries}

\item{b}{Right-hand side vector}
}
\value{
A vector of solutions
}
\description{
This function solves a tridiagonal system of equations of the form
$$Ax = b$$
where the matrix \eqn{A} has the entries \eqn{D_1, D_2, \dots, D_N} on the
main diagonal, \eqn{L_2, L_3, \dots, L_N} on the lower diagonal, and
\eqn{U_1, U_2, \dots, U_{N-1}} on the upper diagonal.
}
\details{
For more details on the double sweep method, see the vignette
\code{vignette("diffusion")}
}
\examples{
U <- c(1, 2, 3, NA)
L <- c(NA, 4, 5, 6)
D <- c(7, 8, 9, 10)
b <- c(11, 12, 13, 14)
solve_double_sweep(U, L, D, b)
}
