% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_params}
\alias{sp_params}
\title{Species parameters for North Sea 14-species model}
\format{
A data frame with 14 rows and columns including species,
Scientific_name, age_mat, l_mat, l_max, a, b, w_mat, w_max, n, p, d,
alpha, and biomass_cutoff.
}
\usage{
sp_params
}
\description{
Species parameters data frame used to build a 14-species North Sea
mizer model.
}
\keyword{datasets}
