% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady.R
\name{steady_diffusion}
\alias{steady_diffusion}
\title{Solve steady-state equation including diffusion}
\usage{
steady_diffusion(params, species)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species for which to solve the steady-state equation.}
}
\value{
A vector of steady-state abundances for the specified species
}
\description{
This function solves the steady-state equation for a single species
including diffusion. It uses an upwind difference scheme to solve the
second-order ODE.
}
\details{
For more details on the upwind difference scheme, see the vignette
\code{vignette("diffusion")}.
}
