% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steadySingleSpecies.R
\name{steadySingleSpecies}
\alias{steadySingleSpecies}
\title{Set initial abundances to single-species steady state abundances}
\usage{
steadySingleSpecies(
  params,
  species = NULL,
  keep = c("egg", "biomass", "number"),
  ...
)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a numeric vector with
the species indices, or a logical vector indicating for each species
whether it is to be selected (TRUE) or not.}

\item{keep}{A string determining which quantity is to be kept constant. The
choices are "egg" which keeps the egg density constant, "biomass" which
keeps the total biomass of the species constant and "number" which keeps
the total number of individuals constant.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This is a thin wrapper around \code{\link[mizer:steadySingleSpecies]{mizer::steadySingleSpecies()}} so that
all \code{mizerEcopath} code uses the implementation provided by the
installed version of \code{mizer}.
}
