% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stomach_data_fit}
\alias{stomach_data_fit}
\title{Stomach content data fit}
\format{
A data frame with 14 rows and columns including species, alpha,
ll, ul, lr, ur, distribution, and min_w_pred.
}
\usage{
stomach_data_fit
}
\description{
Parameters of fitted distributions for stomach content analysis, used in
diet matching.
}
\keyword{datasets}
