% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyTab.R
\name{surveyTab}
\alias{surveyTab}
\alias{surveyTabUI}
\title{Survey tab for tuning gadget}
\usage{
surveyTab(
  input,
  output,
  session,
  params,
  logs,
  age_at_length = NULL,
  catch = NULL,
  ...
)

surveyTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{age_at_length}{A data frame with age-at-length observations for the
selected species. See \code{preprocess_length_at_age()} for expected columns.}

\item{catch}{Data frame holding binned observed catch data. The data can be
binned either into length bins or weight bins. In the former case the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in grams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{gear} (the name of the
gear) and \code{catch} (the number of individuals of a particular species
caught by a particular gear in a size bin).}

\item{...}{Unused}
}
\value{
Invisibly returns \code{NULL}. Used for its side-effect of rendering plots.

UI elements for the survey tab.
}
\description{
Server logic for the survey tab in the tuning gadget. Renders a plot of the
survey number at length and age at length data together with the
corresponding model predictions.
}
\keyword{internal}
