% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tab_name}
\alias{tab_name}
\title{Convert the tab name given by the user to lower case}
\usage{
tab_name(tab)
}
\arguments{
\item{tab}{The tab name}
}
\value{
The tab name in lower case
}
\description{
This function is used to convert the tab name given by the user to lower case,
because the names of the tab functions will always start with lower case.
}
