% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tab_title}
\alias{tab_title}
\title{Return the title for the tab}
\usage{
tab_title(tab)
}
\arguments{
\item{tab}{The tab name}
}
\value{
The tab title
}
\description{
This function is used to return the title for the tab. This is either defined
by the tab author or otherwise is the tab name supplied by the user.
}
