% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tuneParams_add_to_logs}
\alias{tuneParams_add_to_logs}
\title{Add the params object to the logs}
\usage{
tuneParams_add_to_logs(logs, p, params)
}
\arguments{
\item{logs}{The logs object}

\item{p}{The params object}

\item{params}{The reactive params object}
}
\description{
This function is called when the params object needs to be written to the logs.
It updates the params object and writes it to the logs.
}
\keyword{internal}
