% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tuneParams_update_abundance}
\alias{tuneParams_update_abundance}
\title{Update the abundance of a species}
\usage{
tuneParams_update_abundance(p, sp, params, params_old)
}
\arguments{
\item{p}{The params object}

\item{sp}{The species to update}

\item{params}{The reactive params object}

\item{params_old}{The reactive params object before the change}
}
\description{
This function is called when the abundance of a species is changed directly
i.e., not when it is changed as a consequence of a parameter change.
It will make the change permanent by also saving it in params_old.
}
