% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tuneParams_update_params}
\alias{tuneParams_update_params}
\title{Update the params object}
\usage{
tuneParams_update_params(p, params)
}
\arguments{
\item{p}{The params object}

\item{params}{The reactive params object}
}
\description{
This function is called whenever the params object needs to be updated, unless
you also need to write it to the logs, in which case call
\code{tuneParams_add_to_logs()} instead.
It indicates that the params have changed and updates the reactive params object.
}
\keyword{internal}
