% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_params.R
\name{update_params}
\alias{update_params}
\title{Prepare a TMB Objective Function for Optimising Model Parameters}
\usage{
update_params(params, species = 1, pars, data)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species for which to update parameters. By default
the first species in the model.}

\item{pars}{A named list of parameter values as returned by the TMB
optimiser, containing entries for gear selectivity and mortality parameters.}

\item{data}{The data list as returned by \code{\link[=prepare_data]{prepare_data()}}.}
}
\value{
A MizerParams object with updated parameters.
}
\description{
This function returns a list with the data to be passed to the TMB objective
function. The data includes the observed catch data, the model parameters,
and some precomputed values that are used in the likelihood calculation.
The main preprocessing makes sure that we have a comprehensive set of bins
that cover the entire size range, even though there will not be observations
at all sizes. Missing observations should be interpreted as a 0 count.
}
