% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEcopathCatchTotal.R
\name{validEcopathCatch}
\alias{validEcopathCatch}
\title{Validate Ecopath Catch data frame}
\usage{
validEcopathCatch(ecopath_catch, species_params)
}
\arguments{
\item{ecopath_catch}{A data frame with Ecopath catch for each group
as exported by the Ecopath software.}

\item{species_params}{The species parameters data frame}
}
\value{
The validated Ecopath catch data frame
}
\description{
Checks that the Ecopath Catch data frame has the required columns.
}
\details{
The function also deals with the fact that the column names in the Ecopath
data frame can be different from the expected names if it was loaded in with
\code{read.csv} instead of \code{readr::read_csv}.
}
