% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validEcopathParams}
\alias{validEcopathParams}
\title{Validate Ecopath parameter data frame}
\usage{
validEcopathParams(ecopath_params, species_to_groups)
}
\arguments{
\item{ecopath_params}{A data frame with Ecopath parameters for each group
as exported by the Ecopath software.}

\item{species_to_groups}{A named list where the names are mizer species and
the values are vectors of the Ecopath groups/stanzas making up the species.}
}
\value{
The validated Ecopath parameter data frame
}
\description{
Checks that the Ecopath parameter data frame has the required columns and
that the group names are unique and that all groups in the
\code{species_to_groups} list are included in the data frame.
}
\details{
The function also deals with the fact that the column names in the Ecopath
data frame can be different from the expected names if it was loaded in with
\code{read.csv} instead of \code{readr::read_csv}.
}
