% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonBertalanffyGrowth.R
\name{vonBertalanffyGrowth}
\alias{vonBertalanffyGrowth}
\alias{vonBertalanffyGrowth.MizerParams}
\title{von Bertalanffy growth rate}
\usage{
vonBertalanffyGrowth(object)

\method{vonBertalanffyGrowth}{MizerParams}(object)
}
\arguments{
\item{object}{A \code{MizerParams} object.}
}
\value{
An array (species x size) with the von Bertalanffy growth rates.
}
\description{
Return an ArraySpeciesBySize array with the von Bertalanffy growth rates
calculated from the \code{w_inf} and \code{k_vb} species parameters.
}
\details{
The von Bertalanffy growth rate in terms of length \eqn{l} is
\deqn{dl/dt = k_{vb}(L_\inf - l(t)).}
After converting this to weight \eqn{w = al^b} this becomes
\deqn{dw/dt = Aw^n-Bw}
where \eqn{n=1-1/b}, \eqn{A=b\, k_{vb}\, w_{\inf}^{1/b}} and
\eqn{B=b\, k_{vb}}. The von Bertalanffy parameters \eqn{w_{\inf}} and
\eqn{k_{vb}} and the weight-length conversion parameters \eqn{a} and \eqn{b}
need to be provided in the \code{w_inf}, \code{k_vb}, \code{a} and \code{b} columns of the
\code{species_params} data frame contained in the MizerParams object. If \code{w_inf}
is missing then \code{w_max} is used instead and if \code{b} is missing then a default
of \code{b = 3} is used.

The growth rate is evaluated on the size grid \code{w} of the MizerParams object
and is set to zero at sizes above \code{w_inf}, where the von Bertalanffy
expression would otherwise turn negative.
}
\keyword{internal}
