% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceControl.R
\name{abundanceControl}
\alias{abundanceControl}
\alias{abundanceControlUI}
\title{Controlling egg abundance in the tuning gadget}
\usage{
abundanceControl(input, output, session, params, params_old, flags, ...)

abundanceControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\value{
A tagList with a slider for the egg density
}
\description{
This control allows the user to adjust the egg density of the selected
species. Each time the slider is moved, the entire size spectrum for that
species is rescaled so that the initial density matches the new value.
}
\seealso{
Other gadget controls: 
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
