% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceTab.R
\name{abundanceTab}
\alias{abundanceTab}
\alias{abundanceTabUI}
\title{Abundance tab for tuning gadget}
\usage{
abundanceTab(input, ...)

abundanceTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{...}{Unused}

\item{params}{Reactive value holding updated MizerParams object}
}
\description{
The Abundance tab combines the \code{\link[=biomassTab]{biomassTab()}} and the \code{\link[=numberTab]{numberTab()}}
on a single tab, providing a comprehensive view of both biomass and
individual numbers density across size bins and species.
}
\seealso{
Other gadget tabs:
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
