% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceTab.R
\name{abundanceTab}
\alias{abundanceTab}
\alias{abundanceTabUI}
\title{Abundance tab for tuning gadget}
\usage{
abundanceTab(input, ...)

abundanceTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{...}{Unused}

\item{params}{Reactive value holding updated MizerParams object}
}
\description{
The Abundance tab combines the \code{\link[=biomassTab]{biomassTab()}} and the \code{\link[=numberTab]{numberTab()}}
on a single tab, providing a comprehensive view of both biomass and
individual numbers density across size bins and species.
}
\seealso{
Other gadget tabs: 
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
