% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCatch.R
\name{addCatch}
\alias{addCatch}
\title{Sets up gear parameters from scaled catch data}
\usage{
addCatch(params, landings, survey, step)
}
\arguments{
\item{params}{A MizerParams object.}

\item{landings}{A data frame with columns \code{species}, \code{gear}, \code{biomass}.}

\item{survey}{A data frame with columns \code{species}, \code{gear}, \code{biomass}.}

\item{step}{sets effort depending on step, landing effort set to 0 if step =
1 and 1 if step = 2 or 3.}
}
\value{
A MizerParams object with updated gear parameters and effort turned on.
}
\description{
Uses scaled catch data from landings and survey to set up gear parameters
in a MizerParams object. Each gear has a sigmoidal selectivity with 50\% of
individuals selected at maturity weight, and 25\% selected at 90\% of maturity
weight (measured in length).
}
\details{
\code{yield_observed} is set from the data, and \code{catchability} is estimated as
yield divided by observed biomass. This initial estimate is usually too low,
since not all biomass is fully selected by the gear. Use \code{matchCatch()} to
refine catchability after running this function.

Fishing is turned on with an initial effort of 1.
}
