% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassTab.R
\name{biomassTab}
\alias{biomassTab}
\alias{biomassTabUI}
\title{Biomass tab for tuning gadget}
\usage{
biomassTab(input, output, session, params, logs, trigger_update, ...)

biomassTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}
}
\description{
This tab provides a comparison between model biomass and observed
biomass for each species. It features:
\itemize{
\item \strong{Total biomass plot}: A bar-like plot showing circles (model)
and squares (observed) for each species.
\item \strong{Calibrate button}: Rescales the whole model so that the total of
all observed biomasses matches the model total.
\item \strong{Match button}: Adjusts each species' abundance to match its
individual observed biomass.
}
}
\details{
Clicking on a species in the biomass plot selects it. Double-clicking
on a species selects it and scales its abundance to match the clicked
value on the y-axis.
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
