% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassTab.R
\name{biomassTab}
\alias{biomassTab}
\alias{biomassTabUI}
\title{Biomass tab for tuning gadget}
\usage{
biomassTab(input, output, session, params, logs, trigger_update, ...)

biomassTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}
}
\description{
This tab provides a comparison between model biomass and observed
biomass for each species. It features:
\itemize{
\item \strong{Total biomass plot}: A bar-like plot showing circles (model)
and squares (observed) for each species.
\item \strong{Calibrate button}: Rescales the whole model so that the total of
all observed biomasses matches the model total.
\item \strong{Match button}: Adjusts each species' abundance to match its
individual observed biomass.
}
}
\details{
Clicking on a species in the biomass plot selects it. Double-clicking
on a species selects it and scales its abundance to match the clicked
value on the y-axis.
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
