% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchTab.R
\name{catchTab}
\alias{catchTab}
\alias{catchTabUI}
\title{Catch tab for tuning gadget}
\usage{
catchTab(
  input,
  output,
  session,
  params,
  logs,
  trigger_update,
  catch = NULL,
  ...
)

catchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
This tab focuses on comparing model yield with observed catch data.
It includes:
\itemize{
\item \strong{Total yield plot}: Comparison of yearly model yield vs
observed yield for each species.
\item \strong{Catch size distribution}: Plotly visualization of model yield
density vs size, compared with observed binned catch data if
available.
\item \strong{Calibration and Matching}: Buttons to scale either the
entire model or catchability to match observed yields.
}
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
