% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchTab.R
\name{catchTab}
\alias{catchTab}
\alias{catchTabUI}
\title{Catch tab for tuning gadget}
\usage{
catchTab(
  input,
  output,
  session,
  params,
  logs,
  trigger_update,
  catch = NULL,
  ...
)

catchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
This tab focuses on comparing model yield with observed catch data.
It includes:
\itemize{
\item \strong{Total yield plot}: Comparison of yearly model yield vs
observed yield for each species.
\item \strong{Catch size distribution}: Plotly visualization of model yield
density vs size, compared with observed binned catch data if
available.
\item \strong{Calibration and Matching}: Buttons to scale either the
entire model or individual species abundance to match observed yields.
}
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
