% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDiet.R
\name{convertDietMatrix}
\alias{convertDietMatrix}
\title{Convert diet matrix to absolute consumption}
\usage{
convertDietMatrix(diet_matrix, sp, no_sp)
}
\arguments{
\item{diet_matrix}{The diet matrix to convert}

\item{sp}{Species parameters data frame}

\item{no_sp}{Number of species}
}
\value{
The converted diet matrix
}
\description{
This function converts a diet matrix from proportions to absolute consumption
and extracts the part corresponding to species.
}
\keyword{internal}
