% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathAgeTab.R
\name{ecopathAgeTab}
\alias{ecopathAgeTab}
\alias{ecopathAgeTabUI}
\title{Ecopath age tab}
\usage{
ecopathAgeTab(input, output, session, params, logs, age_at_length = NULL, ...)

ecopathAgeTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{age_at_length}{A data frame with age-at-length observations.}

\item{...}{Unused}
}
\description{
This tab displays the age distribution of the selected species. It
allows the user to compare the model's age-at-length distribution
with provided observations (\code{age_at_length}) to verify growth
and cohort dynamics.
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
\keyword{internal}
