% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathAgeTab.R
\name{ecopathAgeTab}
\alias{ecopathAgeTab}
\alias{ecopathAgeTabUI}
\title{Ecopath age tab}
\usage{
ecopathAgeTab(input, output, session, params, logs, age_at_length = NULL, ...)

ecopathAgeTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{age_at_length}{A data frame with age-at-length observations.}

\item{...}{Unused}
}
\description{
Ecopath age tab
}
\keyword{internal}
