% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathCatchTab.R
\name{ecopathCatchTab}
\alias{ecopathCatchTab}
\alias{ecopathCatchTabUI}
\title{Ecopath catch tab}
\usage{
ecopathCatchTab(input, output, session, params, logs, catch = NULL, ...)

ecopathCatchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
This tab provides diagnostic plots for comparison with catch data:
\itemize{
\item \strong{Total yield plot}: A bar-like plot showing circles (model)
and squares (observed) for each species caught by the gear.
\item \strong{Catch size distribution}: Detailed size breakdown of
yield for the selected species and gear.
}
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
