% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathCatchTab.R
\name{ecopathCatchTab}
\alias{ecopathCatchTab}
\alias{ecopathCatchTabUI}
\title{Ecopath catch tab}
\usage{
ecopathCatchTab(input, output, session, params, logs, catch = NULL, ...)

ecopathCatchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
Ecopath catch tab
}
