% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathCatchTab.R
\name{ecopathCatchTab}
\alias{ecopathCatchTab}
\alias{ecopathCatchTabUI}
\title{Ecopath catch tab}
\usage{
ecopathCatchTab(input, output, session, params, logs, catch = NULL, ...)

ecopathCatchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
This tab provides diagnostic plots for comparison with catch data:
\itemize{
\item \strong{Total yield plot}: A bar-like plot showing circles (model)
and squares (observed) for each species caught by the gear.
\item \strong{Catch size distribution}: Detailed size breakdown of
yield for the selected species and gear.
}
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
