% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathDeathTab.R
\name{ecopathDeathTab}
\alias{ecopathDeathTab}
\alias{ecopathDeathTabUI}
\title{Ecopath death tab}
\usage{
ecopathDeathTab(input, output, session, params, logs, diet = NULL, ...)

ecopathDeathTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{diet}{A diet matrix to match the diet of the model to.}

\item{...}{Unused}
}
\description{
This tab provides diagnostic plots for mortality and production. It
allows the user to explore:
\itemize{
\item \strong{Mortality density}: Breakdown of various mortality sources by size
and species.
\item \strong{Production vs Species}: Comparison of biomass produced across
all species in the model.
}
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=simTabUI]{simTabUI()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
