% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathDeathTab.R
\name{ecopathDeathTab}
\alias{ecopathDeathTab}
\alias{ecopathDeathTabUI}
\title{Ecopath death tab}
\usage{
ecopathDeathTab(input, output, session, params, logs, diet = NULL, ...)

ecopathDeathTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{diet}{A diet matrix to match the diet of the model to.}

\item{...}{Unused}
}
\description{
This tab provides diagnostic plots for mortality and production. It
allows the user to explore:
\itemize{
\item \strong{Mortality density}: Breakdown of various mortality sources by size
and species.
\item \strong{Production vs Species}: Comparison of biomass produced across
all species in the model.
}
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
