% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathDietTab.R
\name{ecopathDietTab}
\alias{ecopathDietTab}
\alias{ecopathDietTabUI}
\title{Ecopath diet tab}
\usage{
ecopathDietTab(input, output, session, params, logs, diet = NULL, ...)

ecopathDietTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{diet}{A diet matrix to match the diet of the model to.}

\item{...}{Unused}
}
\description{
Ecopath diet tab
}
