% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathDiffusionControl.R
\name{ecopathDiffusionControl}
\alias{ecopathDiffusionControl}
\alias{ecopathDiffusionControlUI}
\title{Ecopath-specific diffusion control}
\usage{
ecopathDiffusionControl(input, output, session, params, params_old, flags, ...)

ecopathDiffusionControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the exponents
}
\description{
This control adjusts parameters related to age-based diffusion and cohort
dynamics in an Ecopath-aligned mizer model. It includes settings for:
\itemize{
\item \strong{Mean time/concentration of spawning}: Controls the seasonal timing and
duration of reproduction.
\item \strong{Annulus date/age}: Specifies when and at what age annual growth
markers (annuli) are formed.
\item \strong{Diffusion strength}: Adjusts \code{d_over_g}, which determines how much
individuals of the same age can vary in size.
}
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
