% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathExponentControl.R
\name{ecopathExponentControl}
\alias{ecopathExponentControl}
\alias{ecopathExponentControlUI}
\title{Ecopath-specific allometric exponent control}
\usage{
ecopathExponentControl(input, output, session, params, params_old, flags, ...)

ecopathExponentControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the exponents
}
\description{
This control adjusts the allometric exponents for consumption (n),
metabolism (p), and mortality (d). Changing these exponents allows
for fine-tuning how these rates scale with body size.
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
