% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathExponentControl.R
\name{ecopathExponentControl}
\alias{ecopathExponentControl}
\alias{ecopathExponentControlUI}
\title{Ecopath-specific allometric exponent control}
\usage{
ecopathExponentControl(input, output, session, params, params_old, flags, ...)

ecopathExponentControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the exponents
}
\description{
Ecopath-specific allometric exponent control
}
