% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathFishingControl.R
\name{ecopathFishingControl}
\alias{ecopathFishingControl}
\alias{ecopathFishingControlUI}
\title{Ecopath-specific fishing control}
\usage{
ecopathFishingControl(input, output, session, params, params_old, flags, ...)

ecopathFishingControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the gear parameters
}
\description{
Ecopath-specific fishing control
}
