% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathFishingControl.R
\name{ecopathFishingControl}
\alias{ecopathFishingControl}
\alias{ecopathFishingControlUI}
\title{Ecopath-specific fishing control}
\usage{
ecopathFishingControl(input, output, session, params, params_old, flags, ...)

ecopathFishingControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the gear parameters
}
\description{
This control allows for fine-tuning of fishing parameters for each gear
capturing a given species. It allows adjusting:
\itemize{
\item \strong{Catchability}: A global scaling factor for the fishing mortality of
the gear.
\item \strong{Selectivity}: Depending on the selection function (knife-edge,
sigmoid, or double-sigmoid), users can adjust the size at which
individuals become vulnerable to the gear.
}
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
