% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathGrowthControl.R
\name{ecopathGrowthControl}
\alias{ecopathGrowthControl}
\alias{ecopathGrowthControlUI}
\title{Ecopath-specific encounter control}
\usage{
ecopathGrowthControl(input, output, session, params, params_old, flags, ...)

ecopathGrowthControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
Ecopath-specific encounter control
}
