% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathGrowthControl.R
\name{ecopathGrowthControl}
\alias{ecopathGrowthControl}
\alias{ecopathGrowthControlUI}
\title{Ecopath-specific encounter control}
\usage{
ecopathGrowthControl(input, output, session, params, params_old, flags, ...)

ecopathGrowthControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows for manual adjustment of the external encounter rate
coefficient \code{Eiw} for a species. Increasing this value increases the
available food from outside the model, thereby boosting growth without
changing the internal predation dynamics.
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
