% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathGrowthTab.R
\name{ecopathGrowthTab}
\alias{ecopathGrowthTab}
\alias{ecopathGrowthTabUI}
\title{Ecopath growth tab}
\usage{
ecopathGrowthTab(input, output, session, params, logs, size_at_age = NULL, ...)

ecopathGrowthTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{size_at_age}{A data frame with columns 'species', 'age' and 'size'
giving the size of individuals of each species at each age.}

\item{...}{Unused}
}
\description{
Ecopath growth tab
}
