% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathMatchControl.R
\name{ecopathMatchControl}
\alias{ecopathMatchControl}
\alias{ecopathMatchControlUI}
\title{Ecopath-specific match control}
\usage{
ecopathMatchControl(input, output, session, params, params_old, flags, ...)

ecopathMatchControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the matching strength
}
\description{
Ecopath-specific match control
}
