% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathReproductionControl.R
\name{ecopathReproductionControl}
\alias{ecopathReproductionControl}
\alias{ecopathReproductionControlUI}
\title{Ecopath-specific reproduction control}
\usage{
ecopathReproductionControl(
  input,
  output,
  session,
  params,
  params_old,
  flags,
  ...
)

ecopathReproductionControlUI(params, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{The MizerParams object currently being tuned.}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}
}
\value{
A tagList with sliders for the exponents
}
\description{
This control adjusts reproduction parameters using length-based inputs:
\itemize{
\item \strong{l_mat}: Length at 50\% maturity.
\item \strong{l_mat25/l_mat}: Ratio between length at 25\% and 50\% maturity.
\item \strong{l_repro_max}: Maximum length for reproduction.
\item \strong{m}: Allometric exponent for reproductive effort.
}
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
