% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathReproTab.R
\name{ecopathReproTab}
\alias{ecopathReproTab}
\alias{ecopathReproTabUI}
\title{Ecopath reproduction tab}
\usage{
ecopathReproTab(input, output, session, params, logs, ...)

ecopathReproTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Ecopath reproduction tab
}
