% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathSpectraTab.R
\name{ecopathSpectraTab}
\alias{ecopathSpectraTab}
\alias{ecopathSpectraTabUI}
\title{Ecopath spectra tab}
\usage{
ecopathSpectraTab(input, output, session, params, logs, trigger_update, ...)

ecopathSpectraTabUI(params, help = TRUE, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Ecopath spectra tab
}
