% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathSpectraTab.R
\name{ecopathSpectraTab}
\alias{ecopathSpectraTab}
\alias{ecopathSpectraTabUI}
\title{Ecopath spectra tab}
\usage{
ecopathSpectraTab(input, output, session, params, logs, trigger_update, ...)

ecopathSpectraTabUI(params, help = TRUE, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
This tab displays the size spectra of the forground and background
species. It provides controls for:
\itemize{
\item \strong{Scaling the background}: Adjusting the resource and background
species' biomass.
\item \strong{Adjusting/removing background species}: Fine-tuning the
background community to align with the foreground.
}
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
