% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopathSurveyTab.R
\name{ecopathSurveyTab}
\alias{ecopathSurveyTab}
\alias{ecopathSurveyTabUI}
\title{Ecopath survey tab}
\usage{
ecopathSurveyTab(
  input,
  output,
  session,
  params,
  logs,
  age_at_length = NULL,
  catch = NULL,
  ...
)

ecopathSurveyTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{age_at_length}{A data frame with age-at-length observations.}

\item{catch}{Data frame holding binned observed catch data.}

\item{...}{Unused}
}
\description{
Ecopath survey tab
}
\keyword{internal}
