% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getConsumption}
\alias{getConsumption}
\title{Get consumption rate for each species}
\usage{
getConsumption(params, min_w_pred = 0, max_w_pred = Inf)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of consumption rate for each species
}
\description{
For each species, this function returns the rate at which food is consumed
by all individuals of that species. This corresponds to the consumption rate
\eqn{Q} used by Ecopath models.
}
\details{
The consumption is calculated as
\deqn{Q_i = \int_{w_0}^{w_{max}} E_{e.i}(w) (1 - f_i(w)) N_i(w) dw}
where
\eqn{E_{e.i}(w)} is the encounter rate of an individual of species \eqn{i}
and weight \eqn{w} (calculated with \code{\link{getEncounter}}), \eqn{f_i(w)}
is the feeding level of an individual of species \eqn{i} and weight \eqn{w}
(calculated with \code{\link{getFeedingLevel}}) and \eqn{N_i(w)} is the
number density of species \eqn{i} at weight \eqn{w}.
}
\examples{
getConsumption(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
