% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDietMatrix.R
\name{getDietMatrix}
\alias{getDietMatrix}
\title{Get diet matrix}
\usage{
getDietMatrix(params, min_w_pred = 0, max_w_pred = Inf)
}
\arguments{
\item{params}{A MizerParams object}

\item{min_w_pred}{The minimum weight of predators to include in the diet
matrix. Default is 0.}

\item{max_w_pred}{The maximum weight of predators to include in the diet
matrix. Default is \code{Inf}.}
}
\value{
A matrix with dimnames \code{predator} and \code{prey} containing the rates of
biomass flow from a prey species to a predator species (or that part of
the predator species population that is between \code{min_w_pred} and
\code{max_w_pred}).
}
\description{
Returns the diet matrix with one row for each predator species and one
column for each prey species and other ecosystem components. The entries
give the rate at which biomass flows from the prey species to the
predator species.
}
\details{
Note that the diet matrix has the absolute rates of biomass flow, not the
proportions of the diet.

If \code{min_w_pred} or \code{max_w_pred} are set, then the diet matrix only represents
the flow of biomass into the part of the predator species population that has
sizes in the specified range. This may make the returned diet matrix
correspond more closely to diet matrices found in the literature, which are
often estimated from observation of large individuals only or of juveniles
only.
}
\examples{
dimnames(getDietMatrix(NS_params))
getDietMatrix(NS_params)["Cod", ]
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
