% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDietMatrix.R
\name{getDietMatrix}
\alias{getDietMatrix}
\title{Get diet matrix}
\usage{
getDietMatrix(params, min_w_prey = 0, max_w_prey = Inf)
}
\arguments{
\item{params}{A MizerParams object}

\item{min_w_prey}{The minimum weight of prey species to include in the diet
matrix}

\item{max_w_prey}{The maximum weight of prey species to include in the diet
matrix}
}
\value{
A matrix with dimnames \code{predator} and \code{prey} containing the rates of
biomass flow from a prey species to a predator species.
}
\description{
Returns the diet matrix with one row for each predator species and one
column for each prey species and other ecosystem components. The entries
give the rate at which biomass flows from the prey species to the
predator species.
}
\details{
Note that the diet matrix has the absolute rates of biomass flow, not the
proportions of the diet.
}
\examples{
dimnames(getDietMatrix(NS_params))
getDietMatrix(NS_params)["Cod", ]
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
