% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getEcotrophicEfficiency}
\alias{getEcotrophicEfficiency}
\title{Get Ecotrophic Efficiency for each species}
\usage{
getEcotrophicEfficiency(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of ecotrophic efficiency for each species
}
\description{
For each species, returns the ecotrophic efficiency, the proportion of
production that is not lost to mortality external to the model:
\deqn{EE_i = 1-\frac{M0B_i}{P_i}}
where \eqn{M0B_i} is the rate of biomass loss due to external mortality
and \eqn{P_i} is the rate of production.
}
\examples{
getEcotrophicEfficiency(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
