% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getGonadicProduction}
\alias{getGonadicProduction}
\title{Get gonadic production for each species}
\usage{
getGonadicProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of gonadic production for each species
}
\description{
For each species, returns the rate at which gonad biomass is produced by all
individuals of that species.
}
\examples{
getGonadicProduction(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
