% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getGonadicProduction}
\alias{getGonadicProduction}
\title{Get gonadic production for each species}
\usage{
getGonadicProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of gonadic production for each species
}
\description{
For each species, returns the rate at which gonad biomass is produced by all
individuals of that species.
}
\examples{
getGonadicProduction(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
