% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getM0B}
\alias{getM0B}
\title{Get rate at which biomass is lost due to external mortality}
\usage{
getM0B(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of biomass loss rate due to external mortality for
each species
}
\description{
For each species, returns the rate at which biomass is lost due to external
mortality.
}
\details{
This is calculated as
\deqn{M0B_i = \int_{w_0}^{w_{max}} \mu_{ext.i}(w) N_i(w) w dw}
where \eqn{\mu_{ext.i}(w)} is the external mortality rate of an individual
of species \eqn{i} and weight \eqn{w} (obtained with \code{\link{getExtMort}})
and \eqn{N_i(w)} is the number density of species i at weight w.
}
\examples{
getM0B(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
