% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getM2B}
\alias{getM2B}
\title{Get rate at which biomass is lost due to predation mortality}
\usage{
getM2B(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of biomass loss rate due to predation mortality for
each species
}
\description{
For each species, returns the rate at which biomass is lost due to predation
mortality.
}
\details{
This is calculated as
\deqn{M2B_i = \int_{w_0}^{w_{max}} \mu_{p.i}(w) N_i(w) w dw}
where \eqn{\mu_{p.i}(w)} is the predation mortality rate of an individual
of species \eqn{i} and weight \eqn{w} (obtained with \code{\link{getPredMort}})
and \eqn{N_i(w)} is the number density of species i at weight w.
}
\examples{
getM2B(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
