% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMetabolicRespiration}
\alias{getMetabolicRespiration}
\title{Get metabolic respiration for each species}
\usage{
getMetabolicRespiration(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of respiration for each species
}
\description{
For each species, returns the rate at which energy is lost to metabolic respiration
by all individuals of that species. Note that this is different from the
respiration rate \eqn{R} used by Ecopath models, see \code{\link[=getRespiration]{getRespiration()}}.
}
\details{
The respiration rate \eqn{K_i} is calculated as
\deqn{K_i = \int_{w_0}^{w_{max}} k_i(w) N_i(w) dw}
where \eqn{k_i(w)} is the metabolic rate of an individual of species \eqn{i}
and weight \eqn{w} (calculated with \code{\link{getMetabolicRate}}) and
\eqn{N_i(w)} is the number density of species \eqn{i} at weight \eqn{w}.
}
\examples{
getMetabolicRespiration(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
