% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getOffspringProduction}
\alias{getOffspringProduction}
\title{Get rate at which offspring biomass is produced}
\usage{
getOffspringProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of offspring biomass production for each species
}
\description{
For each species, returns the rate at which offspring biomass is produced is
the product of the rate of offspring production (obtained with \code{\link[mizer:getRDD]{mizer::getRDD()}})
and the offspring biomass \eqn{w_{0.i}}.
}
\examples{
getOffspringProduction(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
