% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProduction}
\alias{getProduction}
\title{Get production rate as defined by Ecopath}
\usage{
getProduction(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{
  Arguments passed on to \code{\link[mizer:get_size_range_array]{mizer::get_size_range_array}}
  \describe{
    \item{\code{min_w}}{Smallest weight in size range. Defaults to smallest weight in
the model.}
    \item{\code{max_w}}{Largest weight in size range. Defaults to largest weight in the
model.}
    \item{\code{min_l}}{Smallest length in size range. If supplied, this takes
precedence over \code{min_w}.}
    \item{\code{max_l}}{Largest length in size range. If supplied, this takes precedence
over \code{max_w}.}
  }}
}
\value{
A named vector of production for each species
}
\description{
Unlike \code{\link[=getTotalProduction]{getTotalProduction()}}, this function follows the Ecopath definition
whereby the production rate in a
mass-balanced model is equal to the rate at which biomass is removed by
mortality. Thus the production rate includes the somatic production rate
that is due to growth of individuals and it includes the rate at which
offspring biomass is produced during reproduction. It does not include the
full rate of gonad production because most of that is lost during the
inefficient reproduction process.
The Ecopath production rate is the sum of the somatic production rate
obtained with \code{\link{getSomaticProduction}} and the rate of production
of offspring biomass obtained with \code{\link{getOffspringProduction}}.
}
\examples{
getProduction(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
