% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getProduction}
\alias{getProduction}
\title{Get production rate as defined by Ecopath}
\usage{
getProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of production for each species
}
\description{
Unlike \code{\link[=getTotalProduction]{getTotalProduction()}}, this function follows the Ecopath definition
whereby the production rate in a
mass-balanced model is equal to the rate at which biomass is removed by
mortality. Thus the production rate includes the somatic production rate
that is due to growth of individuals and it includes the rate at which
offspring biomass is produced during reproduction. It does not include the
full rate of gonad production because most of that is lost during the
inefficient reproduction process.
The Ecopath production rate is the sum of the somatic production rate
obtained with \code{\link{getSomaticProduction}} and the rate of production
of offspring biomass obtained with \code{\link{getOffspringProduction}}.
}
\examples{
getProduction(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
