% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getReproductiveLoss}
\alias{getReproductiveLoss}
\title{Get rate of biomass loss due to reproduction}
\usage{
getReproductiveLoss(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of reproduction biomass loss for each species
}
\description{
For each species, returns the difference between the rate of gonad production
(calculated with \link{getGonadicProduction}) and the rate of offspring biomass
production (calculated with \link{getOffspringProduction}).
}
\examples{
getReproductiveLoss(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
