% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getReproductiveLoss}
\alias{getReproductiveLoss}
\title{Get rate of biomass loss due to reproduction}
\usage{
getReproductiveLoss(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{
  Arguments passed on to \code{\link[mizer:get_size_range_array]{mizer::get_size_range_array}}
  \describe{
    \item{\code{min_w}}{Smallest weight in size range. Defaults to smallest weight in
the model.}
    \item{\code{max_w}}{Largest weight in size range. Defaults to largest weight in the
model.}
    \item{\code{min_l}}{Smallest length in size range. If supplied, this takes
precedence over \code{min_w}.}
    \item{\code{max_l}}{Largest length in size range. If supplied, this takes precedence
over \code{max_w}.}
  }}
}
\value{
A named vector of reproduction biomass loss for each species
}
\description{
For each species, returns the difference between the rate of gonad production
(calculated with \link{getGonadicProduction}) and the rate of offspring biomass
production (calculated with \link{getOffspringProduction}).
}
\examples{
getReproductiveLoss(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
