% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getReproductiveLoss}
\alias{getReproductiveLoss}
\title{Get rate of biomass loss due to reproduction}
\usage{
getReproductiveLoss(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of reproduction biomass loss for each species
}
\description{
For each species, returns the difference between the rate of gonad production
(calculated with \link{getGonadicProduction}) and the rate of offspring biomass
production (calculated with \link{getOffspringProduction}).
}
\examples{
getReproductiveLoss(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
