% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResourceInteraction.R
\name{getResourceEncounterRate}
\alias{getResourceEncounterRate}
\title{Get resource encounter rate}
\usage{
getResourceEncounterRate(params)
}
\arguments{
\item{params}{A mizer params object}
}
\value{
A matrix (species x species) of resource encounter rates
}
\description{
This function returns the resource encounter rate for predators.
}
