% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getRespiration}
\alias{getRespiration}
\title{Get respiration rate as defined by Ecopath}
\usage{
getRespiration(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{
  Arguments passed on to \code{\link[mizer:get_size_range_array]{mizer::get_size_range_array}}
  \describe{
    \item{\code{min_w}}{Smallest weight in size range. Defaults to smallest weight in
the model.}
    \item{\code{max_w}}{Largest weight in size range. Defaults to largest weight in the
model.}
    \item{\code{min_l}}{Smallest length in size range. If supplied, this takes
precedence over \code{min_w}.}
    \item{\code{max_l}}{Largest length in size range. If supplied, this takes precedence
over \code{max_w}.}
  }}
}
\value{
A named vector of respiration for each species
}
\description{
Unlike \code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}}, this function follows the Ecopath
definition of respiration as the rate at which assimilated food is NOT used
for production. This is expressed by the first master equation of Ecopath:
\deqn{Q = P + R + U} where \eqn{Q} is the consumption rate, \eqn{P} is the
production rate, \eqn{R} is the respiration rate, and \eqn{U} is the
unassimilated part of the consumption rate, i.e., \eqn{U = (1 - \alpha) Q},
where \eqn{\alpha} is the assimilation efficiency. Solving this for \eqn{R}
we get: \deqn{R = \alpha Q - P}
}
\details{
In mizer, this respiration rate has two components: the metabolic respiration
and the loss due to gonad production that does not result in offspring
biomass. Thus the respiration rate is the sum of the metabolic respiration
rate (calculated with \code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}}) and the rate of
biomass loss due to reproduction (calculated with
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}}).
}
\examples{
getRespiration(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
