% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getRespiration}
\alias{getRespiration}
\title{Get respiration rate as defined by Ecopath}
\usage{
getRespiration(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of respiration for each species
}
\description{
Unlike \code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}}, this function follows the Ecopath
definition of respiration as the rate at which assimilated food is NOT used
for production. This is expressed by the first master equation of Ecopath:
\deqn{Q = P + R + U} where \eqn{Q} is the consumption rate, \eqn{P} is the
production rate, \eqn{R} is the respiration rate, and \eqn{U} is the
unassimilated part of the consumption rate, i.e., \eqn{U = (1 - \alpha) Q},
where \eqn{\alpha} is the assimilation efficiency. Solving this for \eqn{R}
we get: \deqn{R = \alpha Q - P}
}
\details{
In mizer, this respiration rate has two components: the metabolic respiration
and the loss due to gonad production that does not result in offspring
biomass. Thus the respiration rate is the sum of the metabolic respiration
rate (calculated with \code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}}) and the rate of
biomass loss due to reproduction (calculated with
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}}).
}
\examples{
getRespiration(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
