% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSomaticProduction}
\alias{getSomaticProduction}
\title{Get somatic production for each species}
\usage{
getSomaticProduction(params, min_w = 0, max_w = Inf)
}
\arguments{
\item{params}{A MizerParams object}

\item{min_w}{The minimum weight of individuals to include. Default is 0.}

\item{max_w}{The maximum weight of individuals to include. Default is \code{Inf}.}
}
\value{
A named vector of somatic production for each species
}
\description{
For each species, returns the rate at which somatic biomass is produced by all
individuals of that species.
}
\details{
The somatic production rate \eqn{P_{s.i}} is calculated as
\deqn{P_{s.i} = \int_{w_0}^{w_{max}} \phi_i(w) dw}
where \eqn{\phi_i(w)} is the flux in size space for species \eqn{i}
and weight \eqn{w} (calculated with \code{\link{getFlux}}). This includes
both advection and diffusion terms.
}
\examples{
getSomaticProduction(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getTotalProduction]{getTotalProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
