% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSomaticProduction}
\alias{getSomaticProduction}
\title{Get somatic production for each species}
\usage{
getSomaticProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of somatic production for each species
}
\description{
For each species, returns the rate at which somatic biomass is produced by all
individuals of that species.
}
\details{
The somatic production rate \eqn{P_{s.i}} is calculated as
\deqn{P_{s.i} = \int_{w_0}^{w_{max}} g_i(w) N_i(w) dw}
where \eqn{g(w)} is the somatic growth rate of an individual of species \eqn{i}
and weight \eqn{w} (calculated with \code{\link{getEGrowth}})
and \eqn{N_i(w)} is the number density of species \eqn{i} at weight \eqn{w}.
}
\examples{
getSomaticProduction(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()},
\code{\link{getZB}()}
}
\concept{rate functions}
