% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTotalProduction}
\alias{getTotalProduction}
\title{Get total production rate for each species}
\usage{
getTotalProduction(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of gonadic production for each species
}
\description{
For each species, returns the rate at which biomass is produced by all
individuals of that species. This includes both somatic and gonadic
production, which can be obtained individually with \code{\link[=getSomaticProduction]{getSomaticProduction()}}
and \code{\link[=getGonadicProduction]{getGonadicProduction()}}. This differs from the Ecopath definition of
production, see \code{\link[=getProduction]{getProduction()}}.
}
\details{
Used by \code{\link[=matchConsumption]{matchConsumption()}} to calculate how much assimilated energy is left
over after production, in order to estimate the required metabolic loss.
}
\examples{
getTotalProduction(NS_params)
}
\seealso{
Other rate functions:
\code{\link[=getConsumption]{getConsumption()}},
\code{\link[=getDietMatrix]{getDietMatrix()}},
\code{\link[=getEcotrophicEfficiency]{getEcotrophicEfficiency()}},
\code{\link[=getGonadicProduction]{getGonadicProduction()}},
\code{\link[=getM0B]{getM0B()}},
\code{\link[=getM2B]{getM2B()}},
\code{\link[=getMetabolicRespiration]{getMetabolicRespiration()}},
\code{\link[=getOffspringProduction]{getOffspringProduction()}},
\code{\link[=getProduction]{getProduction()}},
\code{\link[=getReproductiveLoss]{getReproductiveLoss()}},
\code{\link[=getRespiration]{getRespiration()}},
\code{\link[=getSomaticProduction]{getSomaticProduction()}},
\code{\link[=getUnassimilated]{getUnassimilated()}},
\code{\link[=getZB]{getZB()}}
}
\concept{rate functions}
