% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getUnassimilated}
\alias{getUnassimilated}
\title{Get unassimilated food}
\usage{
getUnassimilated(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of unassimilated food for each species
}
\description{
For each species, returns the rate at which food is not assimilated.
This is calculated as
\deqn{U_i = (1 - \alpha_i) Q_i}
where \eqn{\alpha_i} is the assimilation efficiency of species \eqn{i}
and \eqn{Q_i} is the consumption rate of species \eqn{i} (calculated with
\code{\link{getConsumption}}).
}
\examples{
getUnassimilated(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getZB}()}
}
\concept{rate functions}
