% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getZB}
\alias{getZB}
\title{Get rate at which biomass is lost due to mortality}
\usage{
getZB(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A named vector of biomass loss rate due to mortality for each species
}
\description{
For each species, returns the rate at which biomass is lost due to mortality.
}
\details{
This is calculated as
\deqn{ZB_i = \int_{w_0}^{w_{max}} \mu_i(w) N_i(w) w dw}
where \eqn{\mu_i(w)} is the mortality rate of an individual of species \eqn{i}
and weight \eqn{w} (calculated with \code{\link{getMort}}) and \eqn{N_i(w)} is the
number density of species \eqn{i} at weight \eqn{w}.
}
\examples{
getZB(NS_params)
}
\seealso{
Other rate functions: 
\code{\link{getConsumption}()},
\code{\link{getDietMatrix}()},
\code{\link{getEcotrophicEfficiency}()},
\code{\link{getGonadicProduction}()},
\code{\link{getM0B}()},
\code{\link{getM2B}()},
\code{\link{getMetabolicRespiration}()},
\code{\link{getOffspringProduction}()},
\code{\link{getProduction}()},
\code{\link{getReproductiveLoss}()},
\code{\link{getRespiration}()},
\code{\link{getSomaticProduction}()},
\code{\link{getTotalProduction}()},
\code{\link{getUnassimilated}()}
}
\concept{rate functions}
