% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthControl.R
\name{growthControl}
\alias{growthControl}
\alias{growthControlUI}
\title{Adjusting predation parameters for growth}
\usage{
growthControl(input, output, session, params, params_old, flags, ...)

growthControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control provides a single slider for the search volume coefficient
\code{gamma}. Each time \code{gamma} is changed, the maximum intake rate coefficient
\code{h} is also adjusted by the same factor. This maintains the same growth
rate at small sizes while allowing for tuning of the overall growth
potential.
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{interactionControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
