% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthTab.R
\name{growthTab}
\alias{growthTab}
\alias{growthTabUI}
\title{Growth tab for tuning gadget
inheritParams biomassTab}
\usage{
growthTab(input, output, session, params, logs, size_at_age = NULL, ...)

growthTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{size_at_age}{A data frame with observed size at age data to be plotted
on top of growth curves. Should contain columns \code{species} (species name as
used in the model), \code{age} (in years) and either \code{weight} (in grams) or
\code{length} (in cm). If both \code{weight} and \code{length} are provided, only \code{weight}
is used.}

\item{...}{Unused}
}
\description{
Growth tab for tuning gadget
inheritParams biomassTab
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
