% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionControl.R
\name{interactionControl}
\alias{interactionControl}
\alias{interactionControlUI}
\title{Controlling the interaction matrix in the tuning gadget}
\usage{
interactionControl(input, output, session, params, params_old, flags, ...)

interactionControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows the user to adjust the interaction strengths between
the selected species and its prey, including the resource:
\itemize{
\item \strong{Resource}: Interaction strength with the resource spectrum.
\item \strong{Prey species}: Drop-down to select a prey species and a slider
to adjust the interaction strength.
\item \strong{Predator species}: Drop-down to select a predator species and
a slider to adjust the interaction strength with that predator.
}
}
\seealso{
Other gadget controls: 
\code{\link{abundanceControl}()},
\code{\link{ecopathDiffusionControl}()},
\code{\link{ecopathExponentControl}()},
\code{\link{ecopathFishingControl}()},
\code{\link{ecopathGrowthControl}()},
\code{\link{ecopathMatchControl}()},
\code{\link{ecopathOtherControl}()},
\code{\link{ecopathReproductionControl}()},
\code{\link{fishingControl}()},
\code{\link{growthControl}()},
\code{\link{otherControl}()},
\code{\link{predationControl}()},
\code{\link{reproductionControl}()},
\code{\link{resourceControl}()}
}
\concept{gadget controls}
