% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_history_fishbase}
\alias{life_history_fishbase}
\title{Life history parameters from FishBase}
\format{
A data frame with columns:
\describe{
\item{species}{Common name of the species}
\item{Scientific_name}{Scientific name}
\item{age_mat}{Age at maturity (years)}
\item{l_mat}{Length at maturity (cm)}
\item{l_max}{Maximum length (cm)}
\item{a}{Length-weight coefficient}
\item{b}{Length-weight exponent}
}
}
\usage{
life_history_fishbase
}
\description{
Life history parameters for Celtic Sea species obtained from FishBase
using \code{\link[=fillDefaultsFromFishBase]{fillDefaultsFromFishBase()}}.
}
\keyword{datasets}
