% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeInteracting.R
\name{makeInteracting}
\alias{makeInteracting}
\title{Enable predator–prey interactions using an interaction matrix}
\usage{
makeInteracting(params, interaction)
}
\arguments{
\item{params}{A MizerParams object}

\item{interaction}{An interaction  matrix where entry \emph{(i, j)} gives the strength
with which species \emph{i} preys on \emph{j}.}
}
\value{
The modified MizerParams object
}
\description{
This function should be called with a non-interacting model (where there is
only external mortality and external encounter) and an interaction matrix.
The external encounter and mortality rates will then be decreased by the
amount of the predation encounter and mortality that is created by that
interaction matrix.
}
\details{
If predation would exceed total mortality or encounter for a species, a warning is issued
and negative values are truncated to zero. This can cause the resulting model to
diverge from the steady state.

Typically used internally by \code{\link[=matchDiet]{matchDiet()}}
}
