% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCatch.R
\name{matchCatch}
\alias{matchCatch}
\title{Match observed catch, yield and production with selectable sigmoid type}
\usage{
matchCatch(
  params,
  species = NULL,
  catch,
  sel_type = c("double", "single"),
  lambda = 2.05,
  yield_lambda = 1,
  production_lambda = 1
)
}
\arguments{
\item{params}{A \code{MizerParams} object}

\item{species}{The species for which to match the catch. Optional. By default
all target species are selected. A vector of species names, or a numeric
vector with the species indices, or a logical vector indicating for each
species whether it is to be selected (TRUE) or not.}

\item{catch}{A data frame containing the observed binned catch data. It must
contain the following columns:
\itemize{
\item \code{length}: The start of each bin.
\item \code{dl}: The width of each bin.
\item \code{count}: The observed count for each bin.
}}

\item{sel_type}{Character; either \code{"double"} (default) for dome-shaped
selectivity, or \code{"single"} for a monotonic logistic.}

\item{lambda}{The slope of the community spectrum. Default is 2.05.}

\item{yield_lambda}{A parameter that controls the strength of the penalty for
deviation from the observed yield.}

\item{production_lambda}{A parameter that controls the strength of the penalty
for deviation from the observed production.}
}
\value{
A MizerParams object with the adjusted external mortality, gear
selectivity, catchability and steady‐state spectrum for the selected
species.
}
\description{
This function adjusts the gear‐selectivity and mortality parameters for one
species so that a steady‐state model reproduces the observed catch size
distribution, the observed yield and the observed production, if available.
}
\examples{
params <- matchCatch(celtic_params, species = "Hake", catch = celtic_catch)
plot_catch(params, species = "Hake", catch = celtic_catch)
}
\seealso{
Other match functions: 
\code{\link{matchConsumption}()},
\code{\link{matchDiet}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchProduction}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
