% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDiet.R
\name{matchDiet}
\alias{matchDiet}
\title{Match the diet matrix of the model to the Ecopath diet matrix}
\usage{
matchDiet(params, diet_matrix, centering = 0, w_prey_cutoff = 1)
}
\arguments{
\item{params}{A MizerParams object}

\item{diet_matrix}{The Ecopath diet matrix as produced by
\code{reduceEcopathDiet()}.}
}
\value{
A MizerParams object with the interaction matrix set so that the
desired diet is achieved.
}
\description{
This function adjusts the interaction matrix of the model so that its diet
matrix matches the Ecopath diet matrix. It then adjusts the external
encounter and mortality rates so that the steady state is not changed.
}
\details{
The function will raise a warning for any species that requires a negative
external mortality rate or negative external encounter rate.
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchConsumption}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchProduction}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
