% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchProduction.R
\name{matchProduction}
\alias{matchProduction}
\alias{matchProductionOnce}
\alias{isProductionMatched}
\title{Match the model's production rate to empirical observations/estimates}
\usage{
matchProduction(params, tol = 0.1, max_iter = 10)

matchProductionOnce(params, steady = TRUE)

isProductionMatched(params, tol = 0.1)
}
\arguments{
\item{params}{A MizerParams object}

\item{tol}{The relative tolerance for matching observed production. Default is 0.1.}

\item{max_iter}{The maximum number of iterations. Default is 10.}

\item{steady}{Logical. If \code{TRUE} (default), returns the model to steady state
after parameter adjustment using \code{\link[=steadySingleSpecies]{steadySingleSpecies()}} and \code{\link[mizer:matchBiomasses]{mizer::matchBiomasses()}}.}
}
\value{
A MizerParams object with the adjusted metabolic and reproductive parameters
to match production.

A MizerParams object with updated parameters and optionally a new steady state.
}
\description{
This function sets the respiration rate according to the first master
equation of Ecopath, which is: \deqn{Q = P + R + U} where \eqn{Q} is the
consumption rate, \eqn{P} is the somatic production rate, \eqn{R} is the
respiration rate, and \eqn{U} is the unassimilated part of the consumption
rate, i.e., \eqn{U = (1 - \alpha) Q}, where \eqn{\alpha} is the assimilation
efficiency. Solving this for \eqn{R} we get: \deqn{R = Q \alpha - P}
}
\details{
Ecopath does not explicitly account for the loss due to gonad production.
However most of the energy invested into gonad production is lost due to
the inefficiency of the reproductive process. Therefore in mizer we need
to include this loss in the respiration rate. Thus
\deqn{R = K + G - R_dd * w_0}
where \eqn{K} is the metabolic rate, \eqn{G} is the gonadic production rate,
\eqn{R_dd} is the rate of offspring production, and \eqn{w_0} is the weight
of an offspring at the size it enters the model. Thus \eqn{G - R_dd * w_0}
is the net loss due to gonad production.

The \code{matchProduction} function calls \code{matchProductionOnce()}
repeatedly until the production matches within the tolerance specified by
the \code{tol} parameter.
The function will return a warning if the maximum number of iterations is
reached without converging.
}
\seealso{
Other match functions: 
\code{\link{matchCatch}()},
\code{\link{matchConsumption}()},
\code{\link{matchDiet}()},
\code{\link{matchExtMortOnce}()},
\code{\link{matchGonadicProportionOnce}()},
\code{\link{matchRespirationOnce}()}
}
\concept{match functions}
