% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerEcopath-package.R
\docType{package}
\name{mizerEcopath-package}
\alias{mizerEcopath}
\alias{mizerEcopath-package}
\title{mizerEcopath: Extending an Ecopath model to a mizer model}
\description{
Tools for creating a mizer model from an Ecopath model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gustavdelius/mizerEcopath}
  \item Report bugs at \url{https://github.com/gustavdelius/mizerEcopath/issues}
}

}
\author{
\strong{Maintainer}: Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]

}
\keyword{internal}
