% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberTab.R
\name{numberTab}
\alias{numberTab}
\alias{numberTabUI}
\title{Number tab for tuning gadget}
\usage{
numberTab(
  input,
  output,
  session,
  params,
  params_old,
  logs,
  trigger_update,
  ...
)

numberTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}
}
\description{
This tab provides a comparison between model numbers (individuals)
and observed numbers for each species. It features:
\itemize{
\item \strong{Total number plot}: Visualization of model total numbers
(circles) vs observed total numbers (squares).
\item \strong{Calibrate button}: Rescales the whole model so that the total
observed numbers match the model total.
\item \strong{Match button}: Adjusts each species' abundance to match its
individual observed number.
}
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
