% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predationControl.R
\name{predationControl}
\alias{predationControl}
\alias{predationControlUI}
\title{Controlling the predation parameters in the tuning gadget}
\usage{
predationControl(input, output, session, params, params_old, flags, ...)

predationControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows the user to adjust parameters that define the
feeding and predation behavior of the selected species:
\itemize{
\item \strong{gamma}: Predation rate coefficient.
\item \strong{h}: Maximum intake rate coefficient.
\item \strong{q}: Allometric exponent for the search volume.
\item \strong{n}: Allometric exponent for the maximum intake rate.
\item \strong{beta}: Preferred predator-prey mass ratio (the peak of the
log-normal predation kernel).
\item \strong{sigma}: The width of the log-normal predation kernel.
}
}
\seealso{
Other gadget controls:
\code{\link[=abundanceControl]{abundanceControl()}},
\code{\link[=ecopathDiffusionControl]{ecopathDiffusionControl()}},
\code{\link[=ecopathExponentControl]{ecopathExponentControl()}},
\code{\link[=ecopathFishingControl]{ecopathFishingControl()}},
\code{\link[=ecopathGrowthControl]{ecopathGrowthControl()}},
\code{\link[=ecopathMatchControl]{ecopathMatchControl()}},
\code{\link[=ecopathOtherControl]{ecopathOtherControl()}},
\code{\link[=ecopathReproductionControl]{ecopathReproductionControl()}},
\code{\link[=fishingControl]{fishingControl()}},
\code{\link[=growthControl]{growthControl()}},
\code{\link[=interactionControl]{interactionControl()}},
\code{\link[=otherControl]{otherControl()}},
\code{\link[=reproductionControl]{reproductionControl()}},
\code{\link[=resourceControl]{resourceControl()}}
}
\concept{gadget controls}
