% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predationControl.R
\name{predationControl}
\alias{predationControl}
\alias{predationControlUI}
\title{Controlling the predation parameters in the tuning gadget}
\usage{
predationControl(input, output, session, params, params_old, flags, ...)

predationControlUI(p, input)
}
\description{
Controlling the predation parameters in the tuning gadget
}
