% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare a TMB Objective Function for Optimizing Model Parameters}
\usage{
prepare_data(
  params,
  species = 1,
  catch,
  yield_lambda = 1,
  production_lambda = 0.01
)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species for which the data is to be prepared. By default
the first species in the model.}

\item{catch}{A data frame containing the observed binned catch data. It must
contain the following columns:
\itemize{
\item \code{length}: The start of each bin.
\item \code{dl}: The width of each bin.
\item \code{count}: The observed count for each bin.
}}

\item{yield_lambda}{A parameter that controls the strength of the penalty for
deviation from the observed yield.}
}
\value{
A list with the data to be passed to the TMB objective function. If
there is no catch data for the species, the function returns NULL.
}
\description{
This function returns a list with the data to be passed to the TMB objective
function. The data includes the observed catch data, the model parameters,
and some precomputed values that are used in the likelihood calculation.
The main preprocessing makes sure that we have a comprehensive set of bins
that cover the entire size range, even though there will not be observations
at all sizes. Missing observations should be interpreted as a 0 count.
}
