% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reduceEcopathDiet}
\alias{reduceEcopathDiet}
\title{Reduce Ecopath diet matrix to mizer species}
\usage{
reduceEcopathDiet(species_params, ecopath_diet)
}
\arguments{
\item{species_params}{A data frame with mizer species parameters}

\item{ecopath_diet}{The Ecopath diet matrix, as exported by the Ecopath
software.}
}
\value{
A matrix with dimnames \code{predator} and \code{prey}, where the row names are
the species names of the predators and the column names are the species
names of the prey and an extra column named "other". The matrix entries
give for each predator the proportion of its diet that comes from each prey
species or from other ecosystem components.
}
\description{
Given an Ecopath diet matrix and a dictionary that maps Ecopath groups to
mizer species, this function returns a diet matrix for the mizer species.
}
