% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproductionControl.R
\name{reproductionControl}
\alias{reproductionControl}
\alias{reproductionControlUI}
\title{Controlling the reproduction parameters in the tuning gadget}
\usage{
reproductionControl(input, output, session, params, params_old, flags, ...)

reproductionControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows the user to adjust parameters that define the maturity
ogive and the maximum size of a species:
\itemize{
\item \strong{w_mat}: Maturity size (50\% maturity).
\item \strong{w_mat25/w_mat}: Ratio between size at 25\% and 50\% maturity,
defining the sharpness of the maturity curve.
\item \strong{w_max}: Maximum size reachable by the species.
\item \strong{m}: Allometric exponent for the reproductive effort.
}
}
\seealso{
Other gadget controls:
\code{\link[=abundanceControl]{abundanceControl()}},
\code{\link[=ecopathDiffusionControl]{ecopathDiffusionControl()}},
\code{\link[=ecopathExponentControl]{ecopathExponentControl()}},
\code{\link[=ecopathFishingControl]{ecopathFishingControl()}},
\code{\link[=ecopathGrowthControl]{ecopathGrowthControl()}},
\code{\link[=ecopathMatchControl]{ecopathMatchControl()}},
\code{\link[=ecopathOtherControl]{ecopathOtherControl()}},
\code{\link[=ecopathReproductionControl]{ecopathReproductionControl()}},
\code{\link[=fishingControl]{fishingControl()}},
\code{\link[=growthControl]{growthControl()}},
\code{\link[=interactionControl]{interactionControl()}},
\code{\link[=otherControl]{otherControl()}},
\code{\link[=predationControl]{predationControl()}},
\code{\link[=resourceControl]{resourceControl()}}
}
\concept{gadget controls}
