% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFeedingLevels.R
\name{setFeedingLevels}
\alias{setFeedingLevels}
\title{Set feeding levels while maintaining constant growth and reproduction}
\usage{
setFeedingLevels(params, f, f_c)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object.}

\item{f}{The target feeding level \eqn{f} for each
species. Must be a vector with one value for each species or a single
value used for all. Defaults to species parameter \code{f0} if exists or else
to 0.6.}

\item{f_c}{The feeding level \eqn{f_c} at which
growth/repro is zero. Defaults to species parameter \code{fc} if exists or else
to 0.2.}
}
\value{
A \linkS4class{MizerParams} object with updated \eqn{h, k_s},
and \code{ext_encounter} rates. The new \eqn{h} and \eqn{k_s} are also
recorded in the given species parameters, so that they are not reverted to
\eqn{h = \infty} and \eqn{k_s = 0} the next time a species parameter is
changed.
}
\description{
This function takes a tuned, non-interacting allometric model (where initially
\eqn{h = \infty} and \eqn{k_s = 0}) and determines the new maximum intake rate
\eqn{h} and metabolic loss \eqn{k_s} parameters. It does this by adjusting
metabolic loss, encounter rate, and maximum intake rate so that the energy
available for reproduction and growth remains unchanged, despite changes in
the feeding level and critical feeding level.
}
\details{
In order to keep the energy for growth and reproduction \eqn{E_r} unchanged,
the function recalculates the encounter rate \eqn{E}, metabolic loss rate
\eqn{k} and maximum intake rate \eqn{h} as follows:

\deqn{E=E_r\frac{f}{\alpha(f-f_c)(1-f)}}
\deqn{h = E_r\frac{1}{\alpha(f-f_c)}}
\deqn{k_s = E_r\frac{f_c}{f-f_c}}

This methodology allows you to set a desired feeding level for each species
without altering the underlying energy balance established during the
initial tuning phase.
}
\section{Mandatory Prerequisites}{

This function will throw an error unless the following conditions are met:
\itemize{
\item \strong{Allometric Scaling}: The model must use allometric rates where the
intake exponent (\eqn{n}) and metabolism exponent (\eqn{p}) are equal.
This ensures the balance remains consistent across all fish sizes.
\item \strong{Initial State}: The species parameters \code{h} and \code{ks}
must have the values \code{h = Inf} and \code{ks = 0}.
\item \strong{Non-interacting Model}: All encounter must be external encounter.
This function is not compatible with models where species interact with
a dynamically realized resource. If interactions exist, call
\code{makeNoninteracting()} first.
}
}

