% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFeedingLevels.R
\name{setFeedingLevels}
\alias{setFeedingLevels}
\title{Set feeding levels while maintaining constant growth and reproduction}
\usage{
setFeedingLevels(params, feeding_level, critical_feeding_level)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object.}

\item{feeding_level}{The target feeding level \eqn{f} for each
species. Must be a vector with one value for each species or a single
value used for all. Defaults to species parameter \code{f0} if exists or else
to 0.6.}

\item{critical_feeding_level}{The feeding level \eqn{f_c} at which
growth/repro is zero. Defaults to species parameter \code{fc} if exists or else
to 0.2.}
}
\value{
A \linkS4class{MizerParams} object with updated \eqn{h, k_s},
and \code{ext_encounter} rates.
}
\description{
This function takes a tuned, non-interacting allometric model (where initially
\eqn{h = \infty} and \eqn{k_s = 0}) and determines the new maximum intake rate
\eqn{h} and metabolic loss \eqn{k_s} parameters. It does this by adjusting
metabolic loss, encounter rate, and maximum intake rate so that the energy
available for reproduction and growth remains unchanged, despite changes in
the feeding level and critical feeding level.
}
\details{
The function follows a three-step methodology to ensure that the consumption
rate does not change in spite of the change of feeding level:
\enumerate{
\item Metabolic Loss Calculation: It uses the critical feeding level \eqn{f_c}
(the level where energy for growth/repro is zero) and the old encounter rate
\eqn{E^{old}} to find the new metabolic loss \eqn{k_{i,w}}:
\deqn{k_{i}(w) = \alpha_i (1 - f_{c,i}) E_{i}^{old}(w)}
\item Encounter Rate Adjustment: It calculates a new encounter rate (\eqn{E_{new}})
required to maintain the original energy available for growth and reproduction
(\eqn{E_{r,i}(w)}) at the target feeding level \eqn{f}:
\deqn{E_{i}^{new}(w) = \frac{E_{r,i}(w) + k_{i}(w)}{\alpha_i (1 - f_i)}}
\item Maximum Intake Back-calculation: Finally, the maximum intake rate \eqn{h}
is back-calculated from the new encounter rate to achieve the desired feeding level:
\deqn{h_{i}(w) = \frac{E_{i}^{new}(w) (1 - f_i)}{f_i}}
}

This methodology allows you to set a desired feeding level for each species
without altering the underlying energy balance established during the
initial tuning phase.
}
\section{Mandatory Prerequisites}{

This function will throw an error unless the following conditions are met:
\itemize{
\item \strong{Allometric Scaling}: The model must use allometric rates where the
intake exponent (\eqn{n}) and metabolism exponent (\eqn{p}) are equal.
This ensures the balance remains consistent across all fish sizes.
\item \strong{Initial State}: The species parameters \code{h} and \code{ks}
must have the values \code{h = Inf} and \code{ks = 0}.
\item \strong{Non-interacting Model}: All encounter must be external encounter.
This function is not compatible with models where species interact with
a dynamically realized resource. If interactions exist, call
\code{makeNoninteracting()} first.
}
}

