% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTab.R
\name{simTabUI}
\alias{simTabUI}
\alias{simTab}
\title{Simulation tab for tuning gadget}
\usage{
simTabUI(...)

simTab(input, output, session, params, params_old, logs, method = "euler", ...)
}
\arguments{
\item{...}{Unused}

\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding the MizerParams object from before
the most recent update.}

\item{logs}{Environment holding the log of steady states.}

\item{method}{The numerical method used for the projection to steady state,
passed on to \code{\link[mizer:project]{mizer::project()}}. Supplied by \code{\link[=tuneParams]{tuneParams()}}.}
}
\description{
This tab shows the results of running the model to steady state. It
displays the biomass dynamics during the simulation, helping to
identify if the system reaches a stable equilibrium.
}
\seealso{
Other gadget tabs:
\code{\link[=abundanceTab]{abundanceTab()}},
\code{\link[=biomassTab]{biomassTab()}},
\code{\link[=catchTab]{catchTab()}},
\code{\link[=deathTab]{deathTab()}},
\code{\link[=dietTab]{dietTab()}},
\code{\link[=ecopathAgeTab]{ecopathAgeTab()}},
\code{\link[=ecopathCatchTab]{ecopathCatchTab()}},
\code{\link[=ecopathDeathTab]{ecopathDeathTab()}},
\code{\link[=ecopathDietTab]{ecopathDietTab()}},
\code{\link[=ecopathGrowthTab]{ecopathGrowthTab()}},
\code{\link[=ecopathReproTab]{ecopathReproTab()}},
\code{\link[=ecopathSpectraTab]{ecopathSpectraTab()}},
\code{\link[=ecopathSurveyTab]{ecopathSurveyTab()}},
\code{\link[=exampleTab]{exampleTab()}},
\code{\link[=growthTab]{growthTab()}},
\code{\link[=numberTab]{numberTab()}},
\code{\link[=ratesTab]{ratesTab()}},
\code{\link[=reproTab]{reproTab()}},
\code{\link[=resourceTab]{resourceTab()}},
\code{\link[=spectraTab]{spectraTab()}},
\code{\link[=stomachTab]{stomachTab()}}
}
\concept{gadget tabs}
