% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTab.R
\name{simTabUI}
\alias{simTabUI}
\alias{simTab}
\title{Simulation tab for tuning gadget}
\usage{
simTabUI(...)

simTab(input, output, session, params, params_old, logs, ...)
}
\arguments{
\item{...}{Unused}

\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}
}
\description{
This tab shows the results of running the model to steady state. It
displays the biomass dynamics during the simulation, helping to
identify if the system reaches a stable equilibrium.
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{spectraTab}()},
\code{\link{stomachTab}()}
}
\concept{gadget tabs}
