% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_params_example}
\alias{species_params_example}
\title{Example species parameters from a few Celtic Sea species.}
\format{
A data frame with 12 rows and 18 columns. Key columns include:
\describe{
\item{species}{Common name of the species}
\item{w_max}{Maximum body weight (g)}
\item{w_mat}{Maturity weight (g)}
\item{a, b}{Length-weight relationship parameters}
\item{n, p, d}{Allometric exponents for consumption, metabolism, and mortality}
\item{alpha}{Assimilation efficiency}
\item{ecopath_groups}{List column mapping Ecopath stanza groups}
}
}
\usage{
species_params_example
}
\description{
A curated subset of species parameters derived from the Celtic Sea model of Lauria et al. (2016),
as well as arbitrary power-law exponents and stanza mapping dictionary.
}
\keyword{datasets}
