% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomachTab.R
\name{stomachTab}
\alias{stomachTab}
\title{Stomach tab for tuning gadget}
\usage{
stomachTab(input, output, session, params, logs, stomach, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{stomach}{Data frame holding observations of prey items in predator
stomachs.}

\item{...}{Unused}
}
\description{
This tab displays the stomach content breakdown for the selected
species, allowing for a comparison between model-predicted
diet and \code{stomach} observations.
}
\seealso{
Other gadget tabs: 
\code{\link{abundanceTab}()},
\code{\link{biomassTab}()},
\code{\link{catchTab}()},
\code{\link{deathTab}()},
\code{\link{dietTab}()},
\code{\link{ecopathAgeTab}()},
\code{\link{ecopathCatchTab}()},
\code{\link{ecopathDeathTab}()},
\code{\link{ecopathDietTab}()},
\code{\link{ecopathGrowthTab}()},
\code{\link{ecopathReproTab}()},
\code{\link{ecopathSpectraTab}()},
\code{\link{ecopathSurveyTab}()},
\code{\link{growthTab}()},
\code{\link{numberTab}()},
\code{\link{ratesTab}()},
\code{\link{reproTab}()},
\code{\link{resourceTab}()},
\code{\link{simTabUI}()},
\code{\link{spectraTab}()}
}
\concept{gadget tabs}
