% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopath_tune.R
\name{tuneEcopath}
\alias{tuneEcopath}
\title{Launch shiny gadget for matching mizer model to Ecopath data}
\usage{
tuneEcopath(
  params,
  catch = NULL,
  diet = NULL,
  controls = c("ecopathGrowth", "ecopathDiffusion", "ecopathFishing",
    "ecopathReproduction", "ecopathOther", "ecopathMatch"),
  tabs = c("ecopathSurvey", "ecopathSpectra", "ecopathCatch", "ecopathGrowth",
    "ecopathRepro", "ecopathDiet", "ecopathDeath"),
  match = c("growth", "yield", "catch", "consumption"),
  preserve = c("erepro", "reproduction_level", "R_max"),
  return_app = FALSE,
  ...
)
}
\arguments{
\item{params}{MizerParams object to tune. If missing, the gadget tries to
recover information from log files left over from aborted previous runs.}

\item{catch}{A data frame containing the observed binned catch data. It must
contain the following columns:
\itemize{
\item \code{species}: The species for which the catch is recorded.
\item \code{gear}: The gear used to collect the catch.
\item \code{length}: The start of each bin.
\item \code{dl}: The width of each bin.
\item \code{catch}: The observed count for each bin.
}}

\item{diet}{An optional diet matrix to display in the Diet tab. If NULL,
the diet tab will not show observed diet data.}

\item{controls}{A character vector of names of input parameter control
sections that should be displayed in the sidebar. See
\code{\link[=tuningGadget]{tuningGadget()}} for details.}

\item{tabs}{A character vector of names of the tabs that should be displayed
in the main section. See \code{\link[=tuningGadget]{tuningGadget()}} for details.}

\item{match}{A character vector. Determines which quantities should be
matched to observations each time the "match" button is pressed.}

\item{preserve}{Specifies whether the \code{reproduction_level} should be
preserved or the maximum reproduction rate \code{R_max} or the reproductive
efficiency \code{erepro} (Default).}

\item{return_app}{Boolean. For testing purposes only.}

\item{...}{Other params needed by individual tabs.}
}
\value{
The tuned MizerParams object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function opens a shiny gadget, an interactive web page. This page has a
side panel with controls for various model parameters and a main panel with
tabs for various diagnostic plots.
}
\details{
This gadget is meant for tuning a model to match Ecopath data.

There is an "Instructions" button near the top left of the gadget that gives
you a quick overview of the user interface.

After you click the "Return" button in the side panel, the function will
return the parameter object in the state at that time, with \code{Rmax} set to
\code{Inf} and \code{erepro} set to the value it had after the last match to Ecopath.
}
