% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{tuneParams}
\alias{tuneParams}
\title{Launch shiny gadget for tuning parameters}
\usage{
tuneParams(
  params,
  controls = c("predation", "fishing", "reproduction", "other", "interaction",
    "resource"),
  tabs = c("Spectra", "Abundance", "Growth", "Repro", "Catch", "Diet", "Death",
    "Resource", "Rates", "Sim"),
  match = c("none"),
  preserve = c("erepro", "reproduction_level", "R_max"),
  return_app = FALSE,
  ...
)
}
\arguments{
\item{params}{MizerParams object to tune. If missing, the gadget tries to
recover information from log files left over from aborted previous runs.}

\item{controls}{A character vector of names of input parameter control
sections that should be displayed in the sidebar.}

\item{tabs}{A character vector of names of the tabs that should be displayed
in the main section.}

\item{match}{A character vector. Determines which quantities should be
matched to observations each time the "steady" button is pressed. Possible
entries are \code{"growth"} (using \code{\link[mizer:matchGrowth]{mizer::matchGrowth()}}), \code{"biomass"} (using
\code{\link[mizer:matchBiomasses]{mizer::matchBiomasses()}}) and \code{"yield"} (using \code{\link[mizer:matchYields]{mizer::matchYields()}}).}

\item{preserve}{Specifies whether the \code{reproduction_level} should be
preserved or the maximum reproduction rate \code{R_max} or the reproductive
efficiency \code{erepro} (Default). See \code{\link[mizer:setBevertonHolt]{mizer::setBevertonHolt()}} for an explanation
of the \code{reproduction_level}.}

\item{return_app}{Boolean. For testing purposes only. When set to TRUE will
return a shinyApp object instead of running the gadget.}

\item{...}{Other params needed by individual tabs.}
}
\value{
The tuned MizerParams object
}
\description{
The function opens a shiny gadget, an interactive web page. This page has a
side panel with controls for various model parameters and a main panel with
tabs for various diagnostic plots.
}
\details{
This gadget is meant for tuning a model to steady state. It is not
meant for tuning the dynamics of the model. That should be done in a second
step using functions like \code{setRmax()} or \code{changeResource()}.

There is an "Instructions" button near the top left of the gadget that gives
you a quick overview of the user interface.

After you click the "Return" button in the side panel, the function will
return the parameter object in the state at that time, with \code{Rmax} set to
\code{Inf} and \code{erepro} set to the value it had after the last run to steady
state.

At any time the gadget allows the user to download the current params object
as an .rds file via the "Download" button.
}
