% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_helpers.R
\name{tuneParams_run_steady}
\alias{tuneParams_run_steady}
\title{Run to steady state}
\usage{
tuneParams_run_steady(
  p,
  params,
  params_old,
  logs,
  session,
  input,
  return_sim = FALSE
)
}
\arguments{
\item{p}{The params object}

\item{params}{The reactive params object}

\item{params_old}{The reactive params object before the change}

\item{logs}{The logs object}

\item{session}{The Shiny session object}

\item{input}{The Shiny input object}

\item{return_sim}{Whether to return the simulation object}
}
\description{
This function is called when the user clicks the "Steady" button in
\code{\link[=tuneParams]{tuneParams()}}. It runs the model to steady state and updates the params
object.
}
