% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_params.R
\name{update_params}
\alias{update_params}
\title{Function that updates params object with new parameter values}
\usage{
update_params(params, species = 1, pars, data, w_select)
}
\arguments{
\item{params}{A MizerParams object to update.}

\item{species}{The species to update (index, name, or logical).}

\item{pars}{A named numeric vector of parameter values from \code{matchCatch}.}

\item{data}{The list of data passed to the objective function.}

\item{w_select}{A logical vector of weight bins used in the likelihood.}
}
\value{
The updated MizerParams object.
}
\description{
This function updates the gear selectivity parameters \code{l50}, \code{l25},
the descending distance \code{d50} from \code{l50} to \code{l50_right}, the catchability,
the steepness \code{U} of the maturity ogive and the coefficient \code{M} of the
power-law mortality rate. It then recalculates the steady state of the model
and rescales it to match the observed biomass.
}
