% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{valid_catch}
\alias{valid_catch}
\title{Validate and extract catch data for a single species}
\usage{
valid_catch(catch, species)
}
\description{
This helper function ensures that the catch data frame contains the required columns,
extracts only the rows for the specified species, and checks that a single gear is used
(current assumption of the code which may subsequently be overhauled).
}
